### 2024-04-26 #6685 @editorialbot
[PRE REVIEW]: Idiomatic Ephemeris Sourcing and Parsing in Julia
**Submitting author:** @cadojo (<a href="http://orcid.org/0000-0001-8655-8125">Joseph D Carpinelli</a>)
**Repository:** https://github.com/cadojo/EphemerisSources.jl
**Branch with paper.md** (empty if default branch): main
**Version:** N/A
