### 2024-04-26 #6685 @dfm
In the meantime, I think we should spend some time getting this submission a little more aligned with the way JOSS reviews typically work.

1. The big issue is that the paper doesn't have the correct syntax. Can you update it to have the syntax [shown in our docs here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), instead of using Quarto? It would be much better if this is on the `main` branch instead of autogenerated and pushed to the `gh-pages` branch, unless there's a strong argument otherwise.

2. Also JOSS typically operates under the assumption that the code and paper live together in the same project, but that isn't the case here. This submission covers 5 separate repositories, and the submitted project only contains documentation. We can work around this, but it does complicate our workflow to some extent. It does look to me like these 5 projects do contain sufficient ["scholarly effort"](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), but each one of the projects is very small (see `cloc` summaries below), with a relatively short revision history. We can definitely work around this, but I'd love to get a comment here from you about the justification for distributing the project like this instead of as a single library.

<summary>Output of cloc on each project</summary>

```bash
~/Desktop
❯ rm -rf main.zip; wget -q https://github.com/cadojo/HorizonsAPI.jl/archive/refs/heads/main.zip; cloc main.zip
      19 text files.
      16 unique files.
       4 files ignored.

Language                     files          blank        comment           code
Julia                            3             84             84            289
Markdown                         4             33              0            119
YAML                             4              4              0            115
Text                             3              2              0             32
TOML                             2              4              0             22
SUM:                            16            127             84            577

~/Desktop
❯ rm -rf main.zip; wget -q https://github.com/cadojo/HorizonsEphemeris.jl/archive/refs/heads/main.zip; cloc main.zip
      16 text files.
      13 unique files.
       4 files ignored.

Language                     files          blank        comment           code
Julia                            3             84             38            256
Markdown                         4             44              0            128
YAML                             4              5              0            126
TOML                             2              4              0             38
SUM:                            13            137             38            548

~/Desktop
❯ rm -rf main.zip; wget -q https://github.com/cadojo/SPICEKernels.jl/archive/refs/heads/main.zip; cloc main.zip
      24 text files.
      21 unique files.
       3 files ignored.

Language                     files          blank        comment           code
Julia                            8           1531           2723            722
YAML                             6              9              0            183
Markdown                         3             58              0            175
TOML                             3              4              0             33
make                             1              0              0              2
SUM:                            21           1602           2723           1115

~/Desktop
❯ rm -rf main.zip; wget -q https://github.com/cadojo/SPICEBodies.jl/archive/refs/heads/main.zip; cloc main.zip
      16 text files.
      13 unique files.
       3 files ignored.

Language                     files          blank        comment           code
YAML                             4              4              0            118
Markdown                         3             33              0            110
Julia                            3             44             78            106
TOML                             3              4              0             28
SUM:                            13             85             78            362

~/Desktop
❯ rm -rf main.zip; wget -q https://github.com/cadojo/SPICEApplications.jl/archive/refs/heads/main.zip; cloc main.zip
      16 text files.
      13 unique files.
       3 files ignored.

Language                     files          blank        comment           code
Julia                            4            168            243            395
YAML                             4              5              0            128
Markdown                         3             17              0             52
TOML                             2              4              0             23
SUM:                            13            194            243            598
```
