+++ src/HorizonsEphemeris.jl
@@ -1,337 +1,0 @@
-"""
-Interact with the JPL HORIZONS REST API.
-
-# Extended Help
-
-## License
-
-$(LICENSE)
-
-## Exports
-
-$(EXPORTS)
-
-## Imports
-
-$(IMPORTS)
-"""
-module HorizonsEphemeris
-
-export NAIF, ephemeris
-
-using DocStringExtensions
-
-@template (FUNCTIONS, METHODS, MACROS) = """
-                                         $(SIGNATURES)
-                                         $(DOCSTRING)
-                                         """
-
-@template (TYPES, CONSTANTS) = """
-                               $(TYPEDEF)
-                               $(DOCSTRING)
-                               """
-
-import JSON
-import HTTP
-
-using CSV
-using Dates
-using Printf
-using SPICE: bodc2n, bodn2c
-using HorizonsAPI: fetch_vectors
-
-"""
-The Horizons API can only process so many discrete time points at once!
-
-# Extended Help
-
-See [#2](https://github.com/cadojo/HorizonsEphemeris.jl/issues/2).
-"""
-const MAX_TLIST_LENGTH = 30
-
-"""
-Given a NAIF ID, return the associated name, if one exists. If the ID provided cannot be
-found, a `KeyError` is thrown.
-"""
-function NAIF(name::AbstractString)::Int
-
-    code = bodn2c(name)
-
-    if !isnothing(code)
-        return code
-    else
-        throw(KeyError(name))
-    end
-
-end
-
-"""
-If a name is given, return the associated NAIF ID, if one exists. If the name provided
-cannot be found, a `KeyError` is thrown.
-"""
-function NAIF(code::Integer)::String
-
-    name = bodc2n(code)
-
-    if !isnothing(name)
-        return name
-    else
-        throw(KeyError(code))
-    end
-
-end
-
-"""
-Parse the body of a `HTTP.Response` for CSV content, and any returned notes about the CSV content.
-"""
-function parse_response(response::HTTP.Response; start="\$\$SOE", stop="\$\$EOE")
-
-    content = JSON.parse(String(response.body))
-
-    result = content["result"]
-
-    if occursin(start, result) && occursin(stop, result)
-        _, ephemeris = split(result, start)
-        ephemeris, notes = split(ephemeris, stop)
-
-        ephemeris = strip(ephemeris)
-
-        notes = join(
-            filter(
-                line -> !all(char -> char == '*', collect(strip(line))),
-                collect(eachline(IOBuffer(notes))),
-            ), "\n"
-        ) |> strip
-
-        return ephemeris, notes
-    else
-        if "error" in keys(content)
-            error("The data delimiters ($start, $stop) were not found in the response! \n\n\t$(content["error"])")
-        else
-            error("The data delimiters ($start, $stop) were not found in the response!")
-        end
-    end
-
-end
-
-"""
-Given any timestamp, return the corresponding modified Julian date.
-"""
-MJD(timestamp) = datetime2julian(DateTime(timestamp)) - 2400000.5
-MJD(timestamp::AbstractFloat) = timestamp - 2400000.5 < zero(timestamp) ? timestamp : timestamp - 2400000.5
-MJD(timestamp::Tuple) = MJD([timestamp...])
-MJD(timestamp::AbstractVector) = map(MJD, timestamp)
-
-"""
-Pull ephemeris data for the provided celestial body, or celestial system.
-"""
-function ephemeris(
[…]
