+++ docs/src/reference/index.md
@@ -1,10 +1,0 @@
-# Reference
-
-_All docstrings!_
-
-```@autodocs
-Modules = [
-    HorizonsEphemeris,
-]
-Order = [:module, :type, :function, :constant]
-```

[…]
