+++ lib/HorizonsAPI/docs/make.jl
@@ -2,20 +2,19 @@ using Documenter
 using HorizonsAPI
 
 makedocs(
-    sitename="HorizonsAPI",
-    format=Documenter.HTML(),
-    modules=[HorizonsAPI],
-    pages=[
+    sitename = "HorizonsAPI",
+    format = Documenter.HTML(),
+    modules = [HorizonsAPI],
+    pages = [
         "Overview" => "index.md",
         "Examples" => "examples/index.md",
         "Reference" => "reference/index.md",
-    ]
+    ],
 )
 
 deploydocs(
-    target="build",
-    repo="github.com/cadojo/HorizonsAPI.jl.git",
-    branch="gh-pages",
-    devbranch="main",
-    versions=["stable" => "v^", "manual", "v#.#", "v#.#.#"],
+    target = "build",
+    branch = "horizons-api",
+    devbranch = "main",
+    versions = ["stable" => "HorizonsAPI-v^", "HorizonsAPI-v#.#", "HorizonsAPI-v#.#.#"],
 )
