+++ lib/HorizonsEphemeris/docs/make.jl
@@ -1,20 +1,23 @@
 using Documenter, HorizonsEphemeris
 
 makedocs(
-    sitename="HorizonsEphemeris",
-    format=Documenter.HTML(),
-    modules=[HorizonsEphemeris],
-    pages=[
+    sitename = "HorizonsEphemeris",
+    format = Documenter.HTML(),
+    modules = [HorizonsEphemeris],
+    pages = [
         "Overview" => "index.md",
         "Examples" => "examples/index.md",
         "Reference" => "reference/index.md",
-    ]
+    ],
 )
 
 deploydocs(
-    target="build",
-    repo="github.com/cadojo/HorizonsEphemeris.jl.git",
-    branch="gh-pages",
-    devbranch="main",
-    versions=["stable" => "v^", "manual", "v#.#", "v#.#.#"],
+    target = "build",
+    branch = "horizons-ephemeris",
+    devbranch = "main",
+    versions = [
+        "stable" => "HorizonsEphemeris-v^",
+        "HorizonsEphemeris-v#.#",
+        "HorizonsEphemeris-v#.#.#",
+    ],
 )
