+++ lib/SPICEKernels/docs/make.jl
@@ -1,19 +1,15 @@
 using Documenter, SPICEKernels
 
 makedocs(
-    sitename="SPICEKernels",
-    format=Documenter.HTML(size_threshold=nothing),
-    modules=[SPICEKernels],
-    pages=[
-        "Overview" => "index.md",
-        "Reference" => "reference/index.md",
-    ]
+    sitename = "SPICEKernels",
+    format = Documenter.HTML(size_threshold = nothing),
+    modules = [SPICEKernels],
+    pages = ["Overview" => "index.md", "Reference" => "reference/index.md"],
 )
 
 deploydocs(
-    target="build",
-    repo="github.com/cadojo/SPICEKernels.jl.git",
-    branch="gh-pages",
-    devbranch="main",
-    versions=["stable" => "v^", "manual", "v#.#", "v#.#.#"],
+    target = "build",
+    branch = "spice-kernels",
+    devbranch = "main",
+    versions = ["stable" => "SPICEKernels-v^", "SPICEKernels-v#.#", "SPICEKernels-v#.#.#"],
 )
