+++ .github/workflows/JOSS.yml
@@ -1,23 +1,0 @@
-on: 
-  push: 
-    branches:
-      - gh-pages
-  workflow_dispatch:
-  
-jobs:
-  paper:
-    runs-on: ubuntu-latest
-    name: JOSS Submission Draft
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@master
-        with:
-          journal: joss
-          paper-path: paper/paper.md
-      - name: Upload
-        uses: actions/upload-artifact@v1
-        with:
-          name: paper
-          path: paper/paper.pdf
