+++ .github/workflows/Documentation.yml
@@ -6,8 +6,6 @@ on:
       - main
     tags: "*"
   workflow_dispatch:
-  schedule:
-    - cron: "0 */6 * * *" # Runs every 6 hours
 
 jobs:
   deploy:
@@ -16,6 +14,58 @@ jobs:
       - uses: actions/checkout@v2
 
       - uses: julia-actions/setup-julia@latest
+
+      #
+      # HorizonsAPI Documentation
+      #
+
+      - name: Install Documentation Dependencies
+        run: julia --project=lib/HorizonsAPI -e 'using Pkg; Pkg.instantiate()'
+
+      - name: Build and deploy
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
+        run: cd lib/HorizonsAPI && julia --project=docs/ docs/make.jl
+
+      #
+      # HorizonsEphemeris Documentation
+      #
+
+      - name: Install Documentation Dependencies
+        run: julia --project=lib/HorizonsEphemeris -e 'using Pkg; Pkg.instantiate()'
+
+      - name: Build and deploy
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
+        run: cd lib/HorizonsEphemeris && julia --project=docs/ docs/make.jl
+
+      #
+      # SPICEKernels Documentation
+      #
+
+      - name: Install Documentation Dependencies
+        run: julia --project=lib/SPICEKernels -e 'using Pkg; Pkg.instantiate()'
+
+      - name: Build and deploy
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
+        run: cd lib/SPICEKernels && julia --project=docs/ docs/make.jl
+
+      #
+      # SPICEBodies Documentation
+      #
+
+      - name: Install Documentation Dependencies
+        run: julia --project=lib/SPICEBodies -e 'using Pkg; Pkg.instantiate()'
+
+      - name: Build and deploy
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }}
+        run: cd lib/SPICEBodies && julia --project=docs/ docs/make.jl
 
       - name: Set up Quarto
         uses: quarto-dev/quarto-actions/setup@v2
@@ -29,16 +79,8 @@ jobs:
       - name: Install Jupyter
         run: pip install jupyter
 
-      - name: Install Paper Dependencies
-        run: julia --project=paper -e 'import Pkg; Pkg.instantiate()'
-
       - name: Install Documentation Dependencies
         run: julia --project=docs -e 'using Pkg; Pkg.instantiate()'
-
-      - name: Render Quarto Project
-        uses: quarto-dev/quarto-actions/render@v2
-        with:
-          path: paper
 
       - name: Build and deploy
         env:

[…]
