+++ lib/HorizonsAPI/.github/workflows/UnitTests.yml
@@ -1,36 +1,0 @@
-name: UnitTests
-on:
-  push:
-    branches:
-      - main
-  schedule:
-    - cron: "0 0 * * 4"
-jobs:
-  test:
-    name:
-      Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{
-      github.event_name }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      fail-fast: true
-      matrix:
-        version:
-          - "1"
-        os:
-          - ubuntu-latest
-        arch:
-          - x64
-    steps:
-      - uses: actions/checkout@v2
-      - uses: julia-actions/setup-julia@v1
-        with:
-          version: ${{ matrix.version }}
-          arch: ${{ matrix.arch }}
-      - uses: actions/cache@v1
-        env:
-          cache-name: cache-artifacts
-        with:
-          path: ~/.julia/artifacts
-          key: ${{ runner.os }}-test-${{ env.cache-name }}-${{hashFiles('**/Project.toml') }}
-      - uses: julia-actions/julia-buildpkg@latest
-      - uses: julia-actions/julia-runtest@latest
