+++ src/EphemerisSources.jl
@@ -1,0 +1,5 @@
+#
+# This file is deliberately empty!
+#
+
+module EphemerisSources end
