+++ Project.toml
@@ -1,0 +1,13 @@
+name = "EphemerisSources"
+uuid = "858900df-f7ae-402a-aa8c-3e4ef85aa271"
+authors = ["Joey Carpinelli <joey@loopy.codes>"]
+version = "0.1.0"
+[deps]
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
