+++ lib/HorizonsAPI/docs/make.jl
@@ -14,6 +14,7 @@ makedocs(
 
 deploydocs(
     target = "build",
+    repo = "github.com/cadojo/EphemerisSources",
     branch = "horizons-api",
     devbranch = "main",
     versions = ["stable" => "HorizonsAPI-v^", "HorizonsAPI-v#.#", "HorizonsAPI-v#.#.#"],
