+++ paper/index.md
@@ -1,0 +1,119 @@
+---
+title: Idiomatic Ephemeris Sourcing and Parsing in Julia
+
+tags:
+  - astrodynamics
+  - astronomy
+  - astrophysics
+  - ephemeris
+  - Julia
+  - SPICE
+  - Horizons
+
+authors:
+  - name: Joseph D Carpinelli
+    orcid: https://orcid.org/0000-0001-8655-8125
+    email: joey@loopy.codes
+
+date: 21 April 2024
+bibliography: references.bib
+---
+
+## Summary
+
+Students and professionals in astronomy, astrodynamics, astrophysics, and other
+related fields often download and parse solar system ephemeris data from two
+major providers:
+[Generic JPL SPICE Kernels](https://naif.jpl.nasa.gov/pub/naif/generic_kernels/),
+and [JPL Horizons](https://ssd.jpl.nasa.gov/horizons/). SPICE Kernels are
+typically read through the SPICE Toolkit, which is available in a variety of
+programming languages, include the C Programming Language with `CSPICE` [@cspice].
+The Julia packages [`CSPICE_jll.jl`](https://github.com/JuliaBinaryWrappers/CSPICE_jll.jl)
+and [`SPICE.jl`](https://github.com/JuliaAstro/SPICE.jl) expose many `CSPICE`
+functions through Julia functions. Julia users can load and interact with SPICE
+kernels through methods such as `SPICE.furnsh` and `SPICE.spkez`.
+Horizons data is available through
+a variety of methods, including email, command-line, graphical web interfaces,
+and a [REST API](https://ssd-api.jpl.nasa.gov/doc/horizons.html) [@horizons].
+
+This paper introduces several packages which allow users to download and process
+ephemeris data idomatically, all from within Julia: `SPICEApplications.jl`,
+`SPICEKernels.jl`, `SPICEBodies.jl`, `HorizonsAPI.jl` and `HorizonsEphemeris.jl`.
+Through the use of these packages, users can share replicatable code which
+automatically fetches publicly available ephemeris data, as opposed to manually
+including ephemeris data files with their source code distribution.
+
+## Statement of Need
+
+While ephemeris users have all of the tools they need to fetch and parse ephemeris
+data within Julia, they do not have the tools to do so _simply_ or
+_idiomatically_. [@sec-need-horizons] and [@sec-need-spice] present the research
+needs filled by each of the five packages introduced in this paper.
+
+### JPL HORIZONS
+
+The two HORIZONS-related packages presented in this paper ---
+[`HorizonsAPI.jl`](https://github.com/cadojo/HorizonsAPI.jl)
+and [`HorizonsEphemeris.jl`](https://github.com/cadojo/HorizonsEphemeris.jl) ---
+are respectively the first Julia packages to precisely match the REST API with
+tab-completion through static keyword arguments, and the first to offer
+automatic response parsing into `NamedTuple` types. The `NamedTuple` output of
+`HorizonsEphemeris.ephemeris`, the top-level method for fetching Cartesian state
+vectors from the HORIZONS platform, allows for easy plotting, file-saving, and
+`DataFrame` construction. Both `HorizonsAPI.jl` and `HorizonsEphemeris.jl` offer
+users a simple, repeatable way to query and parse HORIZONS ephemeris data.
+
+### JPL SPICE
+
+The three SPICE-related packages presented in this paper ---
+[`SPICEApplications.jl`](https://github.com/cadojo/SPICEApplications.jl),
+[`SPICEKernels.jl`](https://github.com/cadojo/SPICEKernels.jl), and
+[`SPICEBodies.jl`](https://github.com/cadojo/SPICEBodies.jl) --- provide
+idiomatic kernel fetching, inspection, and caching from within Julia.
+While SPICE Toolkit executables were _bundled_ in Julia through
[…]
