### 2024-04-27 38163e5
Fixing SPICEBodies docs
