+++ paper/index.md
@@ -13,13 +13,13 @@ tags:
 authors:
   - name: Joseph D Carpinelli
     orcid: https://orcid.org/0000-0001-8655-8125
-    email: joey@loopy.codes
+    email: joseph.d.carpinelli@loopy.codes
 
 date: 21 April 2024
 bibliography: references.bib
 ---
 
-## Summary
+# Summary
 
 Students and professionals in astronomy, astrodynamics, astrophysics, and other
 related fields often download and parse solar system ephemeris data from two
@@ -37,18 +37,24 @@ a variety of methods, including email, command-line, graphical web interfaces,
 and a [REST API](https://ssd-api.jpl.nasa.gov/doc/horizons.html) [@horizons].
 
 This paper introduces several packages which allow users to download and process
-ephemeris data idomatically, all from within Julia: `SPICEApplications.jl`,
+ephemeris data idomatically, all from within Julia:
 `SPICEKernels.jl`, `SPICEBodies.jl`, `HorizonsAPI.jl` and `HorizonsEphemeris.jl`.
 Through the use of these packages, users can share replicatable code which
 automatically fetches publicly available ephemeris data, as opposed to manually
 including ephemeris data files with their source code distribution.
 
-## Statement of Need
+# Statement of Need
 
-While ephemeris users have all of the tools they need to fetch and parse ephemeris
-data within Julia, they do not have the tools to do so _simply_ or
-_idiomatically_. [@sec-need-horizons] and [@sec-need-spice] present the research
-needs filled by each of the five packages introduced in this paper.
+While ephemeris users have the tools they need to fetch and parse ephemeris
+data from multiple sources within Julia, they do not have the tools to do so
+_simply_ or _idiomatically_. Horizons ephemeris data is distributed in plain text with
+surrounding metadata, and manual parsing is required for users to programmatically
+use the fetched ephemeris data. Generic SPICE kernels are freely available, and
+can be used with`CSPICE` (and wrapper libraries) for kernel inspection and data
+retrieval, but new users and students may find the required workflows unfamiliar.
+The packages presented in this paper may be used by students and professionals
+to idiomatically inspect and use ephemeris data, without prior knowledge of SPICE
+Toolkit utilities or REST APIs.
 
 ### JPL HORIZONS
 
@@ -56,27 +62,23 @@ The two HORIZONS-related packages presented in this paper ---
 [`HorizonsAPI.jl`](https://github.com/cadojo/HorizonsAPI.jl)
 and [`HorizonsEphemeris.jl`](https://github.com/cadojo/HorizonsEphemeris.jl) ---
 are respectively the first Julia packages to precisely match the REST API with
-tab-completion through static keyword arguments, and the first to offer
+tab-completion through _static keyword arguments_[^1], and the first to offer
 automatic response parsing into `NamedTuple` types. The `NamedTuple` output of
 `HorizonsEphemeris.ephemeris`, the top-level method for fetching Cartesian state
 vectors from the HORIZONS platform, allows for easy plotting, file-saving, and
 `DataFrame` construction. Both `HorizonsAPI.jl` and `HorizonsEphemeris.jl` offer
 users a simple, repeatable way to query and parse HORIZONS ephemeris data.
 
-### JPL SPICE
+[^1]:
+    The code required to support static keyword arguments was provided by
+    Joseph Wilson, as described in @sec-acknowledgements.
 
-The three SPICE-related packages presented in this paper ---
-[`SPICEApplications.jl`](https://github.com/cadojo/SPICEApplications.jl),
+## JPL SPICE
+
+The two SPICE-related packages presented in this paper ---
[…]
