### 2024-04-27 fd6098d
Fixing SPICEKernels docs
