+++ .vscode/ltex.dictionary.en-US.txt
@@ -1,1 +1,2 @@
 ephemeris
+astrodynamics
