+++ lib/SPICEKernels/docs/src/index.md
@@ -3,10 +3,10 @@
 _Fetch NASA's generic SPICE kernels from within Julia!_
 
 !!! warning
-    This package is not affiliated with or endorsed by NASA, JPL, Caltech, or any other 
+    This package is not affiliated with or endorsed by NASA, JPL, Caltech, or any other
     organization! This is an independently written package by an astrodynamics hobbyist.
 
-## Installation 
+## Installation
 
 ```julia
 pkg> add SPICEKernels # or `Pkg.add("SPICEKernels")`
@@ -15,25 +15,25 @@ pkg> add SPICEKernels # or `Pkg.add("SPICEKernels")`
 ## Motivation
 
 NASA provides generic SPICE kernels with the most common solar system bodies. These kernels,
-which are available for download at <https://naif.jpl.nasa.gov/pub/naif/>, allow you to 
+which are available for download at <https://naif.jpl.nasa.gov/pub/naif/>, allow you to
 use the SPICE Toolkit to query solar system ephemeris data _locally_ in your kernel pool.
-You can fetch position and velocity data for spacecraft and solar system bodies, get shapes 
-and mass parameters for celestial bodies, convert to and from different coordinate frames, 
-and more. But how do you know which kernel to download? And how can you reliably access 
+You can fetch position and velocity data for spacecraft and solar system bodies, get shapes
+and mass parameters for celestial bodies, convert to and from different coordinate frames,
+and more. But how do you know which kernel to download? And how can you reliably access
 each kernel, and ensure your colleagues are downloading _the same_ kernel?
 
 `SPICEKernels.jl` provides a simple interface for downloading, caching, and inspecting
-all of NASA's publicly available Generic Kernels. A continuous integration pipeline is 
+all of NASA's publicly available Generic Kernels. A continuous integration pipeline is
 used to check NASA's HTTP server every week, and any changes to the _names_ of the available
 kernels result in a new patch version of this package.
 The point is, if you're using this package, a simple `Pkg.update()`
-ensures that all of the kernel links are up-to-date and working, and all of the kernel 
+ensures that all of the kernel links are up-to-date and working, and all of the kernel
 docstrings accurately describe their contents. Each kernel can be called like a function,
-i.e. [`de440s()`](@ref); this returns a path to the kernel file, downloading it from NASA's 
-server if necessary. As a result, you can pass each kernel right to `SPICE.jl`'s `furnsh` 
+i.e. [`de440s()`](@ref); this returns a path to the kernel file, downloading it from NASA's
+server if necessary. As a result, you can pass each kernel right to `SPICE.jl`'s `furnsh`
 function for loading into the local kernel pool.
 
-```@example
+```
 using SPICE, SPICEKernels
 
 furnsh(
@@ -46,36 +46,36 @@ furnsh(
 
 ## Features
 
-As mentioned above, each kernel can be _downloaded_, _cached_, and _inspected_ for the 
-expected kernel contents. 
+As mentioned above, each kernel can be _downloaded_, _cached_, and _inspected_ for the
+expected kernel contents.
 
 ### Kernel Fetching
 
-Each kernel is represented as an instance of a [`SPICEKernels.SPICEKernel`](@ref) subtype. All kernel 
+Each kernel is represented as an instance of a [`SPICEKernels.SPICEKernel`](@ref) subtype. All kernel
 types are _callable_, so you can call them like a function to download (if not previously
-cached) the kernel file and return a path to the local file. By default, all kernels are 
-cached to Julia's scratchspace: see [`SPICEKernels.SPICE_KERNEL_DIR`](@ref). You can override this 
-behavior with the `ignorecache` keyword argument. You can copy the fetched ephemeris file 
-to a new location by using the `directory` keyword argument. 
[…]
