### 2024-04-27 #6685 @cadojo
@dfm Thank you for your time and guidance! I think I have addressed all of the topics above, but please let me know if I have any other actions I should take. Also, could you please change the repository to `https://github.com/cadojo/EphemerisSources.jl`?

Thank you again for everything. Please let me know if I can simplify this process further.

---

> 1. The big issue is that the paper doesn't have the correct syntax.

I edited the paper to match the example you referenced. I had thought "Pandoc Markdown" was good enough, but I should be precisely matching the example now. Please let me know if it is still failing to build internally for you all.

> 2. Also JOSS typically operates under the assumption that the code and paper live together in the same project, but that isn't the case here.

Having all of those small projects in separate repos had become a hassle for me as well. I just moved all of those projects into one common repository: [`cadojo/EphemerisSources.jl`](https://github.com/cadojo/EphemerisSources.jl). The `EphemerisSources.jl` package is not actually a library I am releasing, it is just boilerplate to support multiple Julia projects in a single package. So, __I would still like to release this paper for _four_ separate Julia packages__ (I have removed `SPICEApplications`, as that is just a small wrapper and was an outlier in the paper), if that is possible. I've followed some of the [SciML guidance](https://docs.sciml.ai/SciMLStyle/stable/#When-in-doubt,-a-submodule-should-become-a-subpackage-or-separate-package), which recommends sub-packages in a common repo.

So in short, I now have `EphemerisSources.jl` as a common repo for all projects, this paper, and unified documentation. These packages are still released separately with Julia's package manager, but if I understand you correctly, that isn't a problem for this paper
