### 2024-04-27 5a35270
Fixing SPICEKernels docs
