+++ lib/SPICEKernels/Project.toml
@@ -1,7 +1,7 @@
 name = "SPICEKernels"
 uuid = "8e9d28ce-e483-4ef7-bfd9-45b8fef6369c"
-authors = ["Joey Carpinelli <joey@loopy.codes>"]
-version = "1.3.1"
+authors = ["Joey Carpinelli <joseph.d.carpinelli@loopy.codes>"]
+version = "1.4.0"
 
 [deps]
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
