+++ lib/HorizonsEphemeris/Project.toml
@@ -1,7 +1,7 @@
 name = "HorizonsEphemeris"
 uuid = "05ee1981-f730-42d8-b713-4f42d99733dc"
-authors = ["Joe Carpinelli <joseph.d.carpinelli@loopy.software>"]
-version = "0.2.1"
+authors = ["Joe Carpinelli <joseph.d.carpinelli@loopy.codes>"]
+version = "0.3.0"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
