+++ src/EphemerisSources.jl
@@ -1,5 +1,0 @@
-#
-# This file is deliberately empty!
-#
-
-module EphemerisSources end
