+++ test/ephemeris_providers.jl
@@ -1,0 +1,6 @@
+module EphemerisProvidersTests
+
+using Test
+using EphemerisProviders
+
+end
