+++ Project.toml
@@ -1,10 +1,19 @@
-name = "EphemerisSources"
+name = "EphemerisProviders"
 uuid = "858900df-f7ae-402a-aa8c-3e4ef85aa271"
 authors = ["Joey Carpinelli <joey@loopy.codes>"]
 version = "0.1.0"
+
 [deps]
+AstroTime = "c61b5328-d09d-5e37-a9a8-0eb41c39009c"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+DynamicQuantities = "06fc5a27-2a28-4c7c-a15d-362465fb6821"
+HorizonsAPI = "c15253bb-5e94-4b8b-9a02-579bb6c8e3ce"
+HorizonsEphemeris = "05ee1981-f730-42d8-b713-4f42d99733dc"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+RequiredInterfaces = "97f35ef4-7bc5-4ec1-a41a-dcc69c7308c6"
+SPICEBodies = "19f1efb6-7162-484a-b3a1-b52c391689d7"
+SPICEKernels = "8e9d28ce-e483-4ef7-bfd9-45b8fef6369c"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
