### 2024-05-26 c8ce5a5
CompatHelper: add new compat entry for SPICEBodies at version 0.2, (keep existing compat)
