+++ Project.toml
@@ -15,6 +15,9 @@ RequiredInterfaces = "97f35ef4-7bc5-4ec1-a41a-dcc69c7308c6"
 SPICEBodies = "19f1efb6-7162-484a-b3a1-b52c391689d7"
 SPICEKernels = "8e9d28ce-e483-4ef7-bfd9-45b8fef6369c"
 
+[compat]
+SPICEBodies = "0.2"
+
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
