### 2024-05-26 342e009
CompatHelper: add new compat entry for DynamicQuantities at version 0.14, (keep existing compat)
