### 2024-05-26 5f0cd43
CompatHelper: add new compat entry for Reexport at version 1, (keep existing compat)
