### 2024-05-26 2030885
CompatHelper: add new compat entry for SPICEKernels at version 1, (keep existing compat)
