### 2024-05-26 ad32721
CompatHelper: add new compat entry for HorizonsEphemeris at version 0.2, (keep existing compat)
