### 2024-05-26 ad0832f
CompatHelper: add new compat entry for HorizonsAPI at version 0.1, (keep existing compat)
