### 2024-05-26 e6ebf1c
CompatHelper: add new compat entry for RequiredInterfaces at version 0.1, (keep existing compat)
