### 2024-05-26 88e1afd
CompatHelper: add new compat entry for AstroTime at version 0.7, (keep existing compat)
