### 2024-06-07 #6685 @cadojo
Hey @adonath! I will be updating the repository a bit this weekend. All of the code is there for review, but I've run into automated tagging issues after moving code from several packages into this one project.

Also, could you please change the repository to "EphemerisSources.jl"?

https://github.com/cadojo/EphemerisSources.jl
