+++ paper/paper.md
@@ -1,194 +1,0 @@
----
-author:
-- Joseph D Carpinelli
-authors:
-- email: joey@loopy.codes
-  name:
-    family: Carpinelli
-    given: Joseph
-    literal: Joseph D Carpinelli
-  orcid: "https://orcid.org/0000-0001-8655-8125"
-bibliography: references.bib
-clear-hidden-classes: none
-date: April 21, 2024
-date-format: long
-google-scholar: true
-lightbox: auto
-manuscript:
-  article: index.qmd
-  mecaFile: index-meca.zip
-  notebooks:
-  - notebook: index.qmd
-    order: 9999
-    title: Article Notebook
-notebook-preview-options:
-  back: true
-quarto-internal:
-  subarticles:
-  - input: /home/runner/work/ephemeris.loopy.codes/ephemeris.loopy.codes/paper/index.qmd
-    render: true
-    token: nb-article
-remove-hidden: all
-tags:
-- astrodynamics
-- astronomy
-- astrophysics
-- ephemeris
-- Julia
-- SPICE
-- Horizons
-theme: cosmo
-title: Idiomatic Ephemeris Sourcing and Parsing in Julia
-title-block-style: manuscript
-toc-title: Table of contents
-unroll-markdown-cells: true
----
-
-## Summary {#sec-summary}
-
-Students and professionals in astronomy, astrodynamics, astrophysics,
-and other related fields often download and parse solar system ephemeris
-data from two major providers: [Generic JPL SPICE
-Kernels](https://naif.jpl.nasa.gov/pub/naif/generic_kernels/), and [JPL
-Horizons](https://ssd.jpl.nasa.gov/horizons/). SPICE Kernels are
-typically read through the SPICE Toolkit, which is available in a
-variety of programming languages, include the C Programming Language
-with `CSPICE` [@cspice]. The Julia packages
-[`CSPICE_jll.jl`](https://github.com/JuliaBinaryWrappers/CSPICE_jll.jl)
-and [`SPICE.jl`](https://github.com/JuliaAstro/SPICE.jl) expose many
-`CSPICE` functions through Julia functions. Julia users can load and
-interact with SPICE kernels through methods such as `SPICE.furnsh` and
-`SPICE.spkez`. Horizons data is available through a variety of methods,
-including email, command-line, graphical web interfaces, and a [REST
-API](https://ssd-api.jpl.nasa.gov/doc/horizons.html) [@horizons].
-
-This paper introduces several packages which allow users to download and
-process ephemeris data idomatically, all from within Julia:
-`SPICEApplications.jl`, `SPICEKernels.jl`, `SPICEBodies.jl`,
-`HorizonsAPI.jl` and `HorizonsEphemeris.jl`. Through the use of these
-packages, users can share replicatable code which automatically fetches
-publicly available ephemeris data, as opposed to manually including
-ephemeris data files with their source code distribution.
-
-## Statement of Need {#sec-need}
-
-While ephemeris users have all of the tools they need to fetch and parse
-ephemeris data within Julia, they do not have the tools to do so
-*simply* or *idiomatically*.
-[Section 2.1](#sec-need-horizons){.quarto-xref} and
-[Section 2.2](#sec-need-spice){.quarto-xref} present the research needs
-filled by each of the five packages introduced in this paper.
-
-### JPL HORIZONS {#sec-need-horizons}
-
-The two HORIZONS-related packages presented in this paper ---
-[`HorizonsAPI.jl`](https://github.com/cadojo/HorizonsAPI.jl) and
-[`HorizonsEphemeris.jl`](https://github.com/cadojo/HorizonsEphemeris.jl)
---- are respectively the first Julia packages to precisely match the
-REST API with tab-completion through static keyword arguments, and the
[…]
