+++ lib/HorizonsEphemeris/v0.1.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"#HorizonsEphemeris.jl","page":"HorizonsEphemeris.jl","title":"HorizonsEphemeris.jl","text":"","category":"section"},{"location":"","page":"HorizonsEphemeris.jl","title":"HorizonsEphemeris.jl","text":"Solar system ephemeris data for free!","category":"page"},{"location":"","page":"HorizonsEphemeris.jl","title":"HorizonsEphemeris.jl","text":"While HorizonsAPI.jl provides a precise interface which matches the JPL Horizons API, HorizonsEphemeris.jl provides a more user-friendly way to request solar system ephemeris data. This package is in its early stages! The API will stabilize at v1.0. For now,  treat minor version bumps as breaking changes. Patch updates are non-breaking changes.","category":"page"},{"location":"","page":"HorizonsEphemeris.jl","title":"HorizonsEphemeris.jl","text":"warning: Warning\nMore documentation will come, but for now, HorizonsEphemeris only has humble docstrings! For more information, take a look at the project's GitHub repository. There, and in docstrings, you'll find a recurring warning which is reiterated here: this project is not affiliated with or endorsed by NASA, JPL, Caltech, or any other organization!","category":"page"},{"location":"docstrings/#Documentation","page":"Documentation","title":"Documentation","text":"","category":"section"},{"location":"docstrings/","page":"Documentation","title":"Documentation","text":"All docstrings!","category":"page"},{"location":"docstrings/","page":"Documentation","title":"Documentation","text":"Modules = [\n    HorizonsEphemeris,\n]\nOrder = [:module, :type, :function, :constant]","category":"page"},{"location":"docstrings/#HorizonsEphemeris.HorizonsEphemeris","page":"Documentation","title":"HorizonsEphemeris.HorizonsEphemeris","text":"Interact with the JPL HORIZONS REST API.\n\nExtended Help\n\n(Image: Tests) (Image: Docs) (Image: SciML Code Style)\n\nHorizonsEphemeris\n\nA wrapper around the wrapper around JPL's REST API for the HORIZONS solar system ephemeris platform!\n\nWarningThis package is not affiliated with or endorsed by NASA, JPL, Caltech, or any other organization! This is an independently written package by an astrodynamics hobbyist. For more information about code sharing and usage, see the HorizonsEphemeris.jl license file.\n\nInstallation\n\nChoose one of the following two lines!\n\njulia> ]add HorizonsEphemeris\n\njulia> import Pkg; Pkg.add(\"HorizonsEphemeris\");\n\nUsage\n\nAs of v0.1, only Cartesian vectors are supported. You can query the ephemeris function with any solar system body name, or NAIF code; HorizonsEphemeris uses SPICE under the hood to return the appropriate Horizons-compatible NAIF code. After providing the desired solar system body, specify the start time, stop time, and step size for which you want ephemeris data. You'll receive a NamedTuple in return, with keys defaulting to: [:MJD, :Calendar, :X, :Y, :Z, :ΔX, :ΔY, :ΔZ]. The labels for each key can be changed with the header keyword. For example, to get rid of the Unicode character keys, specify header=[:MJD, :Calendar, :X, :Y, :Z, :DX, :DY, :DZ]. This NamedTuple output is automatically compatible with DataFrames. Finally, use the file keyword argument to write the resulting e
[…]
