+++ lib/HorizonsEphemeris/versions.js
@@ -1,2 +1,11 @@
 var DOC_VERSIONS = [
+  "stable",
+  "v0.2",
+  "v0.1",
+  "v0.2.1",
+  "v0.2.0",
+  "v0.1.4",
+  "v0.1.2",
 ];
+var DOCUMENTER_NEWEST = "v0.2.1";
+var DOCUMENTER_STABLE = "stable";
