+++ lib/SPICEKernels/versions.js
@@ -1,17 +1,2 @@
 var DOC_VERSIONS = [
-  "stable",
-  "v1.3",
-  "v1.1",
-  "v1.0",
-  "v0.2",
-  "v0.1",
-  "v1.3.0",
-  "v1.1.8",
-  "v1.1.1",
-  "v1.1.0",
-  "v1.0.0",
-  "v0.2.0",
-  "v0.1.0",
 ];
-var DOCUMENTER_NEWEST = "v1.3.0";
-var DOCUMENTER_STABLE = "stable";
