+++ lib/SPICEKernels/v0.1.0/search_index.js
@@ -1,3 +1,0 @@
-var documenterSearchIndex = {"docs":
-[{"location":"#SPICEKernels.jl","page":"SPICEKernels.jl","title":"SPICEKernels.jl","text":"","category":"section"},{"location":"","page":"SPICEKernels.jl","title":"SPICEKernels.jl","text":"Solar system ephemeris data for free!","category":"page"},{"location":"","page":"SPICEKernels.jl","title":"SPICEKernels.jl","text":"While HorizonsAPI.jl provides a precise interface which matches the JPL Horizons API, SPICEKernels.jl provides an interface for downloading General Kernel files from NASA JPL's public website. More documentation en route!","category":"page"},{"location":"","page":"SPICEKernels.jl","title":"SPICEKernels.jl","text":"warning: Warning\nMore documentation will come, but for now, SPICEKernels only has humble docstrings! For more information, take a look at the project's GitHub repository. There, and in docstrings, you'll find a recurring warning which is reiterated here: this project is not affiliated with or endorsed by NASA, JPL, Caltech, or any other organization!","category":"page"},{"location":"docstrings/#Reference","page":"Reference","title":"Reference","text":"","category":"section"},{"location":"docstrings/","page":"Reference","title":"Reference","text":"All docstrings!","category":"page"},{"location":"docstrings/","page":"Reference","title":"Reference","text":"Modules = [\n    SPICEKernels,\n]\nOrder = [:module, :type, :function, :constant]","category":"page"},{"location":"docstrings/#SPICEKernels.SPICEKernels","page":"Reference","title":"SPICEKernels.SPICEKernels","text":"Download any and all General Kernels provided by NASA JPL!\n\nwarning: Warning\nThis package is not affiliated with or endorsed by NASA, JPL, or any other organization! This is an independently written package by an astrodynamics hobbyist.\n\nExtended Help\n\nImports\n\nBase\nCore\nDates\nDocStringExtensions\nDownloads\nHTTP\nScratch\n\nExports\n\nC_G_1000012_2012_2017\nDSK\nFK\nGravity\nL1_de431\nL2_de431\nL4_de431\nL5_de431\nLSK\nPCK\nSPK\nc2013a1_s105_merged\ncodes_300ast_20100725_fk\ncodes_300ast_20100725_spk\nde430\nde431_part_1\nde431_part_2\nde432s\nde435\nde438\nde440\nde440s\nde441_part_1\nde441_part_2\nde_403_masses\ndss_17_prelim_itrf93_161110\ndss_17_prelim_itrf93_190814_fk\ndss_17_prelim_itrf93_190814_spk\ndss_23_placeholder_itrf93_201017_fk\ndss_23_placeholder_itrf93_201017_spk\ndss_30_itrf93_210201_fk\ndss_30_itrf93_210201_spk\ndss_59_prelim_itrf93_220921_fk\ndss_59_prelim_itrf93_220921_spk\ndss_69_prelim_itrf93_201215_fk\ndss_69_prelim_itrf93_201215_spk\nearth_000101_230910_230618\nearth_200101_990825_predict\nearth_720101_230601\nearth_assoc_itrf93\nearth_fixed\nearth_latest_high_prec\nearth_topo_201023\nearthstns_fx_201023\nearthstns_itrf93_201023\nfetchkernel\ngm_de431\ngm_de440\nhay_a_amica_5_itokawashape_v1_0_512q\nhay_a_amica_5_itokawashape_v1_0_64q\nison\njup344\njup344_nameid\njup344_s2003_j24\njup344_s2003_j24_
[…]
