+++ lib/SPICEBodies/versions.js
@@ -1,8 +1,2 @@
 var DOC_VERSIONS = [
-  "stable",
-  "v0.1",
-  "v0.1.1",
-  "v0.1.0",
 ];
-var DOCUMENTER_NEWEST = "v0.1.1";
-var DOCUMENTER_STABLE = "stable";
