+++ v1.0.0/examples/earth.txt
@@ -1,0 +1,1568 @@
+API VERSION: 1.2
+API SOURCE: NASA/JPL Horizons API
+
+
+
+*******************************************************************************
+Ephemeris / API_USER Sat Jun  8 21:31:09 2024 Pasadena, USA      / Horizons
+*******************************************************************************
+Target body name: Earth (399)                     {source: DE441}
+Center body name: Solar System Barycenter (0)     {source: DE441}
+Center-site name: BODY CENTER
+*******************************************************************************
+Start time      : A.D. 2024-Jan-01 00:00:00.0000 TDB
+Stop  time      : A.D. 2025-Jan-01 00:00:00.0000 TDB
+Step-size       : 1440 minutes
+*******************************************************************************
+Center geodetic : 0.0, 0.0, 0.0                   {E-lon(deg),Lat(deg),Alt(km)}
+Center cylindric: 0.0, 0.0, 0.0                   {E-lon(deg),Dxy(km),Dz(km)}
+Center radii    : (undefined)
+Output units    : KM-S
+Calendar mode   : Mixed Julian/Gregorian
+Output type     : GEOMETRIC cartesian states
+Output format   : 3 (position, velocity, LT, range, range-rate)
+Reference frame : Ecliptic of J2000.0
+*******************************************************************************
+JDTDB
+   X     Y     Z
+   VX    VY    VZ
+   LT    RG    RR
+*******************************************************************************
+$$SOE
+2460310.500000000 = A.D. 2024-Jan-01 00:00:00.0000 TDB 
+ X =-2.600298246493929E+07 Y = 1.445600958366426E+08 Z = 2.324604087693989E+04
+ VX=-2.983302263902714E+01 VY=-5.138481408411204E+00 VZ= 1.105037462794778E-03
+ LT= 4.899394261332376E+02 RG= 1.468801448315927E+08 RR= 2.241843200081927E-01
+2460311.500000000 = A.D. 2024-Jan-02 00:00:00.0000 TDB 
+ X =-2.857666444771557E+07 Y = 1.440936304180347E+08 Z = 2.334921608994901E+04
+ VX=-2.974143119003039E+01 VY=-5.659136117359448E+00 VZ= 1.277159610630907E-03
+ LT= 4.900055493870424E+02 RG= 1.468999680843818E+08 RR= 2.346185734571065E-01
+2460312.500000000 = A.D. 2024-Jan-03 00:00:00.0000 TDB 
+ X =-3.114203859350717E+07 Y = 1.435822325853946E+08 Z = 2.346560445453227E+04
+ VX=-2.964070156155019E+01 VY=-6.178558948963073E+00 VZ= 1.410084996886773E-03
+ LT= 4.900746189757086E+02 RG= 1.469206746261411E+08 RR= 2.446260003623694E-01
+2460313.500000000 = A.D. 2024-Jan-04 00:00:00.0000 TDB 
+ X =-3.369831277226546E+07 Y = 1.430260152768839E+08 Z = 2.359161905796081E+04
[…]
