+++ v1.0.0/examples/apophis.txt
@@ -1,0 +1,201 @@
+API VERSION: 1.2
+API SOURCE: NASA/JPL Horizons API
+
+
+
+*******************************************************************************
+Ephemeris / API_USER Sat Jun  8 21:31:11 2024 Pasadena, USA      / Horizons
+*******************************************************************************
+Target body name: 99942 Apophis (2004 MN4)        {source: JPL#218}
+Center body name: Earth (399)                     {source: DE441}
+Center-site name: BODY CENTER
+*******************************************************************************
+Start time      : A.D. 2022-Jan-01 10:00:00.0000 TDB
+Stop  time      : A.D. 2022-Jan-02 00:00:00.0000 TDB
+Step-size       : 60 minutes
+*******************************************************************************
+Center geodetic : 0.0, 0.0, 0.0                   {E-lon(deg),Lat(deg),Alt(km)}
+Center cylindric: 0.0, 0.0, 0.0                   {E-lon(deg),Dxy(km),Dz(km)}
+Center radii    : 6378.137, 6378.137, 6356.752 km {Equator_a, b, pole_c}
+Keplerian GM    : 3.9860043550702260E+05 km^3/s^2
+Small perturbers: Yes                             {source: SB441-N16}
+Output units    : KM-S, deg, Julian Day Number (Tp)
+Calendar mode   : Mixed Julian/Gregorian
+Output type     : GEOMETRIC osculating elements
+Output format   : 10
+EOP file        : eop.240607.p240831
+EOP coverage    : DATA-BASED 1962-JAN-20 TO 2024-JUN-07. PREDICTS-> 2024-AUG-30
+Reference frame : Ecliptic of J2000.0
+*******************************************************************************
+Initial IAU76/J2000 heliocentric ecliptic osculating elements (au, days, deg.):
+  EPOCH=  2459215.5 ! 2021-Jan-01.00 (TDB)         Residual RMS= .17952        
+   EC= .1915216892529348   QR= .7458270479403062   TP= 2459101.039422418       
+   OM= 204.0389271089114   W=  126.6520518833791   IN= 3.336751319526047       
+  Equivalent ICRF heliocentric cartesian coordinates (au, au/d):
+   X=-4.098530837957945E-01  Y= 9.070198867765817E-01  Z= 3.267919471926086E-01
+  VX=-1.507552412223498E-02 VY=-3.713559924880165E-03 VZ=-1.761867984500986E-03
+Asteroid physical parameters (km, seconds, rotational period in hours):        
+   GM= n.a.                RAD= .170               ROTPER= 30.56               
+   H= 19.09                G= .240                 B-V= n.a.                   
+Non-default Earth gravity model (EOBL_LIM= au):                                
+   EOBL_LIM= 1.000                                 EOBL_MOD = 2x0 (J2 only)    
+Asteroid non-gravitational force model (AMRAT= m^2/kg;A1,A2,A3=au/d^2;R0=au):  
+   AMRAT=  0.                                                                  
+   A1= 4.999999873689E-13  A2= -2.901092102547E-14 A3= 0.                      
+ Non-standard or simulated/proxy model:                                        
[…]
