+++ paper/_quarto.yml
@@ -10,9 +10,20 @@ format:
   html:
     include-before-body: 
       text: |
-        :::{.callout-important appearance="simple"}
-        This page is a **pre-print**! The published paper is available at 
-        the [Journal of Open Source Software](https://joss.theoj.org). 
+        <div class="callout-warning callout callout-style-simple">
+            <div class="callout-body d-flex">
+                <div class="callout-icon-container">
+                    <i class="callout-icon"></i>
+                    </div>
+                    <div class="callout-body-container">
+                    <p><i>
+                      This page is a **pre-print**! The published paper is available at 
+                      the Journal of Open Source Software.
+                    </i></p>
+                </div>
+            </div>
+        </div>
+
         :::
 
     comments:
