+++ v1.0.0/examples/earth.txt
@@ -4,7 +4,7 @@ API SOURCE: NASA/JPL Horizons API
 
 
 *******************************************************************************
-Ephemeris / API_USER Sat Jun  8 21:31:09 2024 Pasadena, USA      / Horizons
+Ephemeris / API_USER Sat Jun  8 21:47:23 2024 Pasadena, USA      / Horizons
 *******************************************************************************
 Target body name: Earth (399)                     {source: DE441}
 Center body name: Solar System Barycenter (0)     {source: DE441}
