+++ README.md
@@ -1,3 +1,6 @@
+[![Tests](https://github.com/cadojo/EphemerisSources.jl/actions/workflows/Tests.yml/badge.svg)](https://github.com/cadojo/EphemerisSources.jl/actions/workflows/Tests.yml)
+[![Documentation](https://github.com/cadojo/EphemerisSources.jl/actions/workflows/Tests.yml/badge.svg)](https://github.com/cadojo/EphemerisSources.jl/actions/workflows/Documentation.yml)
+
 # 🪐 `EphemerisSources.jl`
 
 > [!IMPORTANT]
