### 2024-06-09 b0b20cc
Updating paper HTML theme
