### 2024-06-09 381227f
Removing Paper.yml
