### 2024-06-09 28207c0
Removing Paper from build
