+++ paper/post.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
 cp ./references.bib ./_manuscript
+mv ./_manuscript paper
+mkdir _manuscript
+mv paper _manuscript
