+++ paper/post.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-cp ./references.bib ./_manuscript
-mv ./_manuscript paper
-mkdir _manuscript
-mv paper _manuscript
