+++ paper/_quarto.yml
@@ -1,10 +1,13 @@
 project:
   type: manuscript
-  output-dir: _manuscript
-  post-render: ./post.sh
+  output-dir: paper
+
+  
 
 manuscript:
   article: paper.md
+  resources: 
+    - "*.bib"
 
 format:
   html:
@@ -23,7 +26,7 @@ format:
                     <div class="callout-body-container">
                     <p><i>
                       This page is a <b>pre-print</b>! The published paper is available at 
-                      the <a jref="https://joss.theoj.org">Journal of Open Source Software</a>.
+                      the <a href="https://joss.theoj.org">Journal of Open Source Software</a>.
                     </i></p>
                 </div>
             </div>
