+++ paper/post.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+mv _manuscript paper
+mkdir _manuscript
+mv paper _manuscript
