+++ paper/_quarto.yml
@@ -1,5 +1,6 @@
 project:
   type: manuscript
+  output-dir: _manuscript
   post-render: ./post.sh
 
 manuscript:
