+++ paper/_quarto.yml
@@ -12,9 +12,9 @@ format:
   html:
     title-block-style: manuscript
     title-block-banner-color: white
-    title-block-banner: "#474b63"
+    title-block-banner: "#313549"
     include-in-header: 
-      text: <meta theme-color content="#474b63"/>
+      text: <meta theme-color content="#313549"/>
     include-before-body: 
       text: |
         <div class="callout-note callout callout-style-simple">
@@ -41,7 +41,7 @@ format:
         - _style/common.scss
         - _style/mocha.scss
 
-    css: _style/static.css
+    css: static.css
 
     mainfont: "'Noto Sans', sans-serif"
     monofont: "'Noto Sans Mono', monospace;"
