### 2024-06-10 60c2c23
Adds back paper rendering
