### 2024-06-10 ef029db
Fixing paper rendering
