+++ lib/SPICEKernels/dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"reference/#Reference","page":"Reference","title":"Reference","text":"","category":"section"},{"location":"reference/","page":"Reference","title":"Reference","text":"All docstrings!","category":"page"},{"location":"reference/","page":"Reference","title":"Reference","text":"Modules = [\n    SPICEKernels,\n]\nOrder = [:module, :type, :function, :constant]","category":"page"},{"location":"reference/#SPICEKernels.SPICEKernels","page":"Reference","title":"SPICEKernels.SPICEKernels","text":"Download any and all Generic Kernels provided by NASA JPL!\n\nwarning: Warning\nThis package is not affiliated with or endorsed by NASA, JPL, or any other organization! This is an independently written package by an astrodynamics hobbyist.\n\nExtended Help\n\nImports\n\nBase\nCore\nDates\nDocStringExtensions\nDownloads\nHTTP\nScratch\n\nExports\n\nC_G_1000012_2012_2017\nDSK\nFK\nGravity\nL1_de431\nL2_de431\nL4_de431\nL5_de431\nLSK\nPCK\nSPK\nc2013a1_s105_merged\ncodes_300ast_20100725_fk\ncodes_300ast_20100725_spk\nde430\nde431_part_1\nde431_part_2\nde432s\nde435\nde438\nde440\nde440s\nde441_part_1\nde441_part_2\nde_403_masses\ndss_17_prelim_itrf93_161110\ndss_17_prelim_itrf93_190814_fk\ndss_17_prelim_itrf93_190814_spk\ndss_23_placeholder_itrf93_201017_fk\ndss_23_placeholder_itrf93_201017_spk\ndss_30_itrf93_210201_fk\ndss_30_itrf93_210201_spk\ndss_59_prelim_itrf93_220921_fk\ndss_59_prelim_itrf93_220921_spk\ndss_69_prelim_itrf93_201215_fk\ndss_69_prelim_itrf93_201215_spk\nearth_000101_240901_240608\nearth_200101_990825_predict\nearth_720101_230601\nearth_assoc_itrf93\nearth_fixed\nearth_latest_high_prec\nearth_topo_201023\nearthstns_fx_201023\nearthstns_itrf93_201023\nfetchkernel\ngm_de431\ngm_de440\nhay_a_amica_5_itokawashape_v1_0_512q\nhay_a_amica_5_itokawashape_v1_0_64q\nison\njup344\njup344_nameid\njup344_s2003_j24\njup344_s2003_j24_nameid\njup346\njup346_nameid\njup365\nkernel\nlatest_leapseconds_lsk\nmar097\nmars_iau2000_v1\nmoon_080317\nmoon_assoc_me\nmoon_assoc_pa\nmoon_de440_220930\nmoon_pa_de403_1950_2198\nmoon_pa_de418_1950_2050\nmoon_pa_de421_1900_2050\nmoon_pa_de440_200625\nnaif0011\nnaif0012_lsk\nndosl_140530_v01_fk\nndosl_140530_v01_spk\nndosl_190716_v01_fk\nndosl_190716_v01_spk\nndosl_190716_v02_fk\nndosl_190716_v02_spk\nnear_a_msi_5_erosshape_v1_0_512q\nnear_a_msi_5_erosshape_v1_0_64q\nnep095\nnep095_nameid\nnep097\nnep097xl_801\nnep097xl_899\nnep101xl\nnep101xl_802\nnep103\nnep103_nameid\npck00010\npck00011\npck00011_n0066\nphobos512\nphobos_2014_09_22\nphobos_3_3\nplu06
[…]
