### 2024-06-10 2d900ab
Darkens banner color
