### 2024-06-10 56af701
Removes justification
