### 2024-06-19 #6685 @cadojo
Yes! Thank you, please let me know if I can clarify anything. I noticed a couple typos in the paper, and I will fix them next week as I'm traveling this week.
