### 2024-06-21 #6914 @adonath
@cadojo, @tamasgal and  @andrew-saydjari welcome to the review thread for your paper. All future communication regarding this submission will take place here. Thank you for your contribution and for agreeing to participate in the review process!

Review instructions are provided at the top of this page. We strongly encourage reviewers to submit issues and pull requests on the software repository. When doing so, please reference openjournals/joss-reviews#REVIEW_NUMBER to create a link to this thread, allowing me to monitor progress. Feel free to post comments, questions, or suggestions in this thread as they arise during your review. In our experience, sharing feedback incrementally is more effective than waiting until you've examined the entire package.

Our target timeframe for completing reviews is 3-4 weeks. If you need additional time, please let me know. We can also utilize the editorial boot to set up automatic reminders if you know you'll be away for a known period of time.

If you have any questions or concerns, don't hesitate to contact me (@adonath) directly.
