### 2024-07-06 JuliaAstro/EphemerisSources.jl#12 @andrew-saydjari
Paper Comments
I think this paper helps readers navigate and understand the packages included in the repo and provides context about some of the existing packages. My comments are mostly targeted at clarity about functionality and improving these comparisons.

- [x] clarify exactly what ephemeris data is. I think all of the packages here are just exposing the (x,y,z,vx,vy,vz) state vector at a given point or range in time. It would be good to spell that out for the non-experts.
- [x] SPICE.jl and HORIZONS.jl seem to clearly represent the prior state of the art in Julia. Having tried to use both of them, I agree that they are harder to use than the package(s) you have developed. It might be helpful to more clearly state that they provided lower level expert interface and you are providing a easier access interface, with reduced functionality (and be clear about what functionality is lost). For example, I think you are dropping obs_tbl support for Horizons? Might be good to tell people which use cases they need to go elsewhere for (and where to go).
- [x] Line 24: withCSPICE -> with CSPICE
- [x] Line 41: I suggest inserting "Previously" at the beginning of the sentence (rewording it appropriately).
- [x] Line 41: aGeneric -> Generic
- [x] Line 46: insert "runs" at the end of the line
- [x] Line 47: multiple times daily, (delete the comma)
- [x] I was confused by what you meant by "exports the kernels as variables to Julia," but I think you mean that it generates a PR to keep the repo up to date so that all kernels are variables users can query.
- [x] Why is SPICEApplications.jl mentioned in the paper, seems to be in the call stack, but is not part of this meta repo?
- [x] #26
- [x] With respect to the possibility of adding Ephemerides.jl support, a sentence saying what benefit that would bring might be nice.
