### 2024-07-06 JuliaAstro/EphemerisSources.jl#14 @andrew-saydjari
HorizonsEphemeris Comments
- [x] The example uses earth.X and earth.Y, but you have to use earth.x and earth.y.
- [x] The compat for HorizonsEphemeris requires HorizonsAPI = "0.1" which is limits it to any version <v1.0.0, but HorizonsAPI is not v1. It would be good to update this compat, and possibly add automated CompatHelper checking to help track things like this.
- [x] I briefly tried to pass ephemeris an AstroTime epoch. I failed. If it is possible, documenting it would be helpful. If not, this might be a feature request (which you should not view as acceptance blocking), but that would be really nice.
- [x] I *think* that all of the names you can pass are listed here (https://naif.jpl.nasa.gov/pub/naif/). If so, it would be good to be explicit about the list of "bodies" that ephemeris can take and document that clearly (ideally both in overview and in the local page).
- [x] Default units of ephemeris appear to be AU-D, rather than KM-S which is what I would have expected based on https://ssd-api.jpl.nasa.gov/doc/horizons.html. If you have made a choice to have a default that is inconsistent with the default specified there, I think it is necessary to document that.
