### 2024-07-06 JuliaAstro/EphemerisSources.jl#15 @andrew-saydjari
SPICEBodies Comments
- [x] (Probably a feature request) It would be much better if the result of the KernelBody type on an epoch was a named tuple, which could have DataFrames support like HorizonsEphemeris, rather that a Vector{Vector{}} because that is fairly opaque. I think it would not take *that* much effort to do this, but I will also view this comment as not being an acceptance-blocker.
