### 2024-07-06 JuliaAstro/EphemerisSources.jl#16 @andrew-saydjari
High-Level Comments
- [x] #37
- [x] You have a nice statement of need on your https://ephemeris.loopy.codes/lib/SPICEKernels/dev/ page, but not on the overview landing page. Adding one on the overview page would be helpful.
- [x] I did not see any community guidelines (which is one of our review checklist items). Please add something that addresses that requirement to help people know how willing you are to interact with them in terms of contributions, issues, and support.
- [x] A single, common, real-world example might be a good addition to the docs. If you showed how to solve it with both the SPICE and Horizons sub packages, that would be amazing. I am trying to understand the best real-world example to use here. I see you have https://github.com/cadojo/GeneralAstrodynamics.jl, so maybe you have an example motivated by that work? Personally, I would care about querying an ephemeris most in the context of knowing how far away an object would be from where my telescope is pointing... but that does not seem to be an easy application/goal of these packages, which are more focused on cartesian coordinates and state vectors. Maybe a "real-world" application is to calculate the closest approach of two objects? What is the minimum distance between Gaia and JWST over the last year? (that is a question I would find at least intriguing)
- [x] You never show an example that uses the velocity information... seems like cooking up and example that uses velocities might be nice.
