### 2024-07-06 #6914 @andrew-saydjari
I have completed my first round of review and left comments on the paper (cadojo/EphemerisSources.jl#12), overall documentation (cadojo/EphemerisSources.jl#13), the Horizons subpackages (cadojo/EphemerisSources.jl#14), SPICE subpackages (cadojo/EphemerisSources.jl#15), and some high-level comments (cadojo/EphemerisSources.jl#16).
