### 2024-07-14 1b0f0ce
Updates SPICEBodies to a slightly new API
