+++ lib/SPICEBodies/Project.toml
@@ -1,12 +1,13 @@
 name = "SPICEBodies"
 uuid = "19f1efb6-7162-484a-b3a1-b52c391689d7"
 authors = ["Joey Carpinelli <joseph.d.carpinelli@loopy.codes>"]
-version = "1.0.0"
+version = "1.1.0"
 
 [deps]
 AstroTime = "c61b5328-d09d-5e37-a9a8-0eb41c39009c"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 SPICE = "5bab7191-041a-5c2e-a744-024b9c3a5062"
 
 [compat]
