+++ lib/EphemerisSourcesBase/docs/src/reference/index.md
@@ -1,0 +1,10 @@
+# Reference
+
+_All docstrings!_
+
+```@autodocs
+Modules = [
+    EphemerisSourcesBase,
+]
+Order = [:module, :type, :function, :constant]
+```
