+++ lib/EphemerisSourcesBase/README.md
@@ -1,0 +1,14 @@
+[![Tests](https://github.com/cadojo/EphemerisSourcesBase.jl/workflows/Tests/badge.svg)](https://github.com/cadojo/EphemerisSourcesBase.jl/actions?query=workflow%3ATests)
+[![Docs](https://github.com/cadojo/EphemerisSourcesBase.jl/workflows/Documentation/badge.svg)](https://cadojo.github.io/EphemerisSourcesBase.jl)
+[![SciML Code Style](https://img.shields.io/static/v1?label=Style&message=SciML&color=9668e2&labelColor=3E474F)](https://github.com/SciML/SciMLStyle)
+
+# `EphemerisSourcesBase`
+
+_Common types and methods for `EphemerisSources.jl`!_
+
+> **Warning**
+>
+> This package is not affiliated with or endorsed by NASA, JPL, Caltech, or any
+> other organization! This is an independently written package by an
+> astrodynamics hobbyist. See the `EphemerisSourcesBase.jl` [license file](./LICENSE) for
+> more information about code sharing and usage!
