+++ lib/HorizonsEphemeris/Project.toml
@@ -18,7 +18,7 @@ SPICEApplications = "e12e0822-0612-48d4-a4bb-92984dd6b6ec"
 CSV = "0.10"
 DocStringExtensions = "0.9"
 HTTP = "^1.8"
-HorizonsAPI = "0.1"
+HorizonsAPI = "1"
 JSON = "^0.21"
 SPICE = "0.2"
 SPICEApplications = "0.3"
