+++ .github/workflows/TagBot.yml
@@ -28,11 +28,12 @@ jobs:
     if: github.event_name == 'workflow_dispatch' || github.actor == 'JuliaTagBot'
     runs-on: ubuntu-latest
     steps:
-      - name: Tag EphemerisSources.jl
+      - name: Tag EphemerisSourcesBase.jl
         uses: JuliaRegistries/TagBot@v1
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
           ssh: ${{ secrets.DOCUMENTER_KEY }}
+          subdir: lib/EphemerisSourcesBase
       - name: Tag HorizonsAPI.jl
         uses: JuliaRegistries/TagBot@v1
         with:
