+++ Project.toml
@@ -5,7 +5,6 @@ version = "0.1.0"
 
 [deps]
 AstroTime = "c61b5328-d09d-5e37-a9a8-0eb41c39009c"
-EphemerisSourcesBase = "b8f1d712-e88d-4f2a-8498-e66ce688ee29"
 HorizonsAPI = "c15253bb-5e94-4b8b-9a02-579bb6c8e3ce"
 HorizonsEphemeris = "05ee1981-f730-42d8-b713-4f42d99733dc"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
