+++ lib/EphemerisSourcesBase/docs/src/index.md
@@ -3,11 +3,9 @@
 _Common types and methods for `EphemerisSources.jl`!_
 
 > **Warning**
->
 > This package is not affiliated with or endorsed by NASA, JPL, Caltech, or any
 > other organization! This is an independently written package by an
-> astrodynamics hobbyist. See the `EphemerisSourcesBase.jl` [license file](./LICENSE) for
-> more information about code sharing and usage!
+> astrodynamics hobbyist.
 
 ## Overview
 
