+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"reference/#Reference","page":"Reference","title":"Reference","text":"","category":"section"},{"location":"reference/","page":"Reference","title":"Reference","text":"All docstrings!","category":"page"},{"location":"reference/","page":"Reference","title":"Reference","text":"Modules = [\n    EphemerisSourcesBase,\n]\nOrder = [:module, :type, :function, :constant]","category":"page"},{"location":"reference/#EphemerisSourcesBase.EphemerisSourcesBase","page":"Reference","title":"EphemerisSourcesBase.EphemerisSourcesBase","text":"Common types and methods for EphemerisSources.jl.\n\nExtended Help\n\nImports\n\nAstroTime\nBase\nCore\nDocStringExtensions\nSPICE\n\nExports\n\n\n\n\n\n","category":"module"},{"location":"reference/#EphemerisSourcesBase.CelestialBody","page":"Reference","title":"EphemerisSourcesBase.CelestialBody","text":"abstract type CelestialBody\n\nAn abstract supertype for celestial bodies in ephemeris data.\n\n\n\n\n\n","category":"type"},{"location":"reference/#EphemerisSourcesBase.naifcode-Tuple{EphemerisSourcesBase.CelestialBody}","page":"Reference","title":"EphemerisSourcesBase.naifcode","text":"naifcode(body)\n\n\nGiven a CelestialBody, a NAIF ID code, or a body name, return the corresponding NAIF ID.\n\n\n\n\n\n","category":"method"},{"location":"#EphemerisSourcesBase","page":"Overview","title":"EphemerisSourcesBase","text":"","category":"section"},{"location":"","page":"Overview","title":"Overview","text":"Common types and methods for EphemerisSources.jl!","category":"page"},{"location":"","page":"Overview","title":"Overview","text":"Warning This package is not affiliated with or endorsed by NASA, JPL, Caltech, or any other organization! This is an independently written package by an astrodynamics hobbyist.","category":"page"},{"location":"#Overview","page":"Overview","title":"Overview","text":"","category":"section"},{"location":"","page":"Overview","title":"Overview","text":"Coming soon.","category":"page"}]
+}
