### 2024-07-15 792ee39
CompatHelper: add new compat entry for SPICE at version 0.2, (keep existing compat)
