### 2024-07-15 aa78e6d
CompatHelper: add new compat entry for SPICEBodies at version 1, (keep existing compat)
