### 2024-07-15 39adc99
CompatHelper: add new compat entry for HorizonsAPI at version 1, (keep existing compat)
