### 2024-07-15 a3ee46c
CompatHelper: add new compat entry for SPICEKernels at version 1, (keep existing compat)
