### 2024-07-15 9bc84e9
CompatHelper: add new compat entry for HorizonsEphemeris at version 1, (keep existing compat)
