### 2024-07-16 bcbcaf0
Adds compat entries for EphemerisSourcesBase
