+++ lib/EphemerisSourcesBase/Project.toml
@@ -12,6 +12,10 @@ SPICEApplications = "e12e0822-0612-48d4-a4bb-92984dd6b6ec"
 
 [compat]
 julia = "1.6"
+AstroTime = "0.7"
+DocStringExtensions = "0.9"
+SPICE = "0.2"
+SPICEApplications = "0.3"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
