### 2024-07-17 f218a78
CompatHelper: add new compat entry for AstroTime at version 0.7, (keep existing compat)
