### 2024-07-17 4545f6f
CompatHelper: add new compat entry for Reexport at version 1, (keep existing compat)
