### 2024-07-20 c4b8959
Adds compat entries
