+++ Project.toml
@@ -12,6 +12,16 @@ SPICE = "5bab7191-041a-5c2e-a744-024b9c3a5062"
 SPICEBodies = "19f1efb6-7162-484a-b3a1-b52c391689d7"
 SPICEKernels = "8e9d28ce-e483-4ef7-bfd9-45b8fef6369c"
 
+[compat]
+julia = "1.6"
+AstroTime = "0.7"
+HorizonsAPI = "1"
+HorizonsEphemeris = "1"
+SPICEKernels = "1.4"
+SPICEBodies = "1.1"
+SPICE = "0.2"
+Reexport = "1.2"
+
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
