### 2024-07-21 37381c4
Fixes SPICEBodies doc error
