+++ lib/SPICEBodies/Project.toml
@@ -1,7 +1,7 @@
 name = "SPICEBodies"
 uuid = "19f1efb6-7162-484a-b3a1-b52c391689d7"
 authors = ["Joey Carpinelli <joseph.d.carpinelli@loopy.codes>"]
-version = "1.1.1"
+version = "1.1.2"
 
 [deps]
 AstroTime = "c61b5328-d09d-5e37-a9a8-0eb41c39009c"
