### 2024-07-30 JuliaAstro/EphemerisSources.jl#13 @andrew-saydjari
That sounds just fine to me. I saw the progress you made on #12 and am glad you got integrated with JuliaAstro. We completely understand you only being able to work on it periodically.
