### 2024-07-31 b1b185f
Moves documentation to JuliaAstro
