+++ paper/paper.md
@@ -67,8 +67,8 @@ Toolkit utilities or REST APIs.
 ### JPL HORIZONS
 
 The two HORIZONS-related packages presented in this paper ---
-[`HorizonsAPI.jl`](https://github.com/cadojo/EphemerisSources.jl/tree/main/lib/HorizonsAPI)
-and [`HorizonsEphemeris.jl`](https://github.com/cadojo/EphemerisSources.jl/tree/main/lib/HorizonsEphemeris) ---
+[`HorizonsAPI.jl`](https://github.com/JuliaAstro/EphemerisSources.jl/tree/main/lib/HorizonsAPI)
+and [`HorizonsEphemeris.jl`](https://github.com/JuliaAstro/EphemerisSources.jl/tree/main/lib/HorizonsEphemeris) ---
 are respectively the first Julia packages to precisely match the REST API with
 tab-completion through _static keyword arguments_[^1], and the first to offer
 automatic response parsing into `NamedTuple` types. The `NamedTuple` output of
@@ -84,8 +84,8 @@ users a simple, repeatable way to query and parse HORIZONS ephemeris data.
 ## JPL SPICE
 
 The two SPICE-related packages presented in this paper ---
-[`SPICEKernels.jl`](https://github.com/cadojo/EphemerisSources.jl/tree/main/lib/SPICEKernels), and
-[`SPICEBodies.jl`](https://github.com/cadojo/EphemerisSources.jl/tree/main/lib/SPICEBodies) --- provide
+[`SPICEKernels.jl`](https://github.com/JuliaAstro/EphemerisSources.jl/tree/main/lib/SPICEKernels), and
+[`SPICEBodies.jl`](https://github.com/JuliaAstro/EphemerisSources.jl/tree/main/lib/SPICEBodies) --- provide
 idiomatic kernel fetching, inspection, and caching from within Julia.
 Julia users interact with SPICE kernels by downloading publicly-available
 a[Generic Kernels](https://naif.jpl.nasa.gov/pub/naif/generic_kernels/), and
@@ -95,10 +95,10 @@ their chosen application, and that they know how to use CSPICE functions to
 retrieve their desired data. `SPICEKernels.jl` and `SPICEBodies.jl` offer
 idiomatic interfaces to ephemeris fetching and parsing parsing respectively.
 Continuous integration in the
-[`SPICEKernels.jl` repository](https://github.com/cadojo/EphemerisSources.jl/tree/main/.github/workflows/SPICE.yml)
+[`SPICEKernels.jl` repository](https://github.com/JuliaAstro/EphemerisSources.jl/tree/main/.github/workflows/SPICE.yml)
 multiple times daily, and automatically exports all available generic kernels
 as variables in Julia. SPICE Toolkit executables (provided by
-[`SPICEApplications.jl`](https://github.com/cadojo/SPICEApplications.jl))
+[`SPICEApplications.jl`](https://github.com/JuliaAstro/SPICEApplications.jl))
 are used to retrieve a description of each kernel's
 contents, and place that description in the Julia variable's docstring. As a
 result, users can use tab-completion and Julia's built-in documentation tools
@@ -110,7 +110,7 @@ a provided instance in time.
 # Usage
 
 For detailed usage examples for, consult the common
-[documentation site](https://ephemeris.loopy.codes).
+[documentation site](https://juliaastro.github.io/docs/EphemerisSources.jl).
 
 # External Packages
 
