+++ lib/SPICEKernels/docs/make.jl
@@ -10,7 +10,7 @@ makedocs(
 deploydocs(
     target = "build",
     branch = "spice-kernels",
-    repo = "github.com/cadojo/EphemerisSources.jl",
+    repo = "github.com/JuliaAstro/EphemerisSources.jl",
     devbranch = "main",
     tag_prefix = "SPICEKernels-",
     versions = ["stable" => "SPICEKernels-v^", "SPICEKernels-v#.#", "SPICEKernels-v#.#.#"],
