+++ lib/EphemerisSourcesBase/README.md
@@ -1,5 +1,5 @@
-[![Tests](https://github.com/cadojo/EphemerisSourcesBase.jl/workflows/Tests/badge.svg)](https://github.com/cadojo/EphemerisSourcesBase.jl/actions?query=workflow%3ATests)
-[![Docs](https://github.com/cadojo/EphemerisSourcesBase.jl/workflows/Documentation/badge.svg)](https://cadojo.github.io/EphemerisSourcesBase.jl)
+[![Tests](https://github.com/JuliaAstro/EphemerisSourcesBase.jl/workflows/Tests/badge.svg)](https://github.com/JuliaAstro/EphemerisSourcesBase.jl/actions?query=workflow%3ATests)
+[![Docs](https://github.com/JuliaAstro/EphemerisSourcesBase.jl/workflows/Documentation/badge.svg)](https://JuliaAstro.github.io/EphemerisSourcesBase.jl)
 [![SciML Code Style](https://img.shields.io/static/v1?label=Style&message=SciML&color=9668e2&labelColor=3E474F)](https://github.com/SciML/SciMLStyle)
 
 # `EphemerisSourcesBase`

[…]
