+++ README.md
@@ -1,5 +1,5 @@
 [![Tests](https://github.com/JuliaAstro/EphemerisSources.jl/actions/workflows/Tests.yml/badge.svg)](https://github.com/JuliaAstro/EphemerisSources.jl/actions/workflows/Tests.yml)
-[![Documentation](https://github.com/JuliaAstro/EphemerisSources.jl/actions/workflows/Documentation.yml/badge.svg)](https://github.com/JuliaAstro/EphemerisSources.jl/actions/workflows/Documentation.yml)
+[![Documentation](https://github.com/JuliaAstro/EphemerisSources.jl/actions/workflows/Documentation.yml/badge.svg)](https://juliaastro.org/EphemerisSources.jl/stable)
 
 # 🪐 `EphemerisSources.jl`
 
@@ -9,31 +9,40 @@
 > endorsed by NASA, JPL, Caltech, or any other organization! They are
 > independently written packages by an astrodynamics hobbyist.
 
+## Installation
+
+Choose one of the following, or manually install one or more of the sub-packages
+below.
+
+```julia
+pkg> add EphemerisSources
+```
+
+```julia
+import Pkg
+Pkg.add("EphemerisSources")
+```
+
 ## Packages
 
 This repository includes the top-level documentation for several solar system
 ephemeris tools, all written with the Julia Programming Language. To learn
 more about how to use these tools, consult the
-[documentation website](https://juliaastro.github.io/docs/EphemerisSources.jl), or visit each package
-repository.
+[documentation website](https://juliaastro.org/EphemerisSources.jl).
 
-> [!NOTE]
->
-> At this time, there is no package named `EphemerisSources.jl`. This repository
-> hosts all of the packages shown below.
-
-| Package                                         | Description                                                                                                  |
-| ----------------------------------------------- | ------------------------------------------------------------------------------------------------------------ |
-| [`HorizonsAPI.jl`](lib/HorizonsAPI)             | A precise JPL HORIZONS REST API client implementation, with keyword arguments for each acceptable parameter. |
-| [`HorizonsEphemeris.jl`](lib/HorizonsEphemeris) | Convenience wrappers around the JPL HORIZONS REST API.                                                       |
-| [`SPICEKernels.jl`](lib/SPICEKernels)           | All [generic kernels](https://naif.jpl.nasa.gov/naif/data_generic.html) exported as variable constants.      |
-| [`SPICEBodies.jl`](lib/SPICEBodies)             | Idiomatic wrappers around SPICE and [`SPICE.jl`](https://github.com/JuliaAstro/SPICE.jl) methods.            |
+| Package                                         | Description                                                                                                                           |
+| ----------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------- |
+| [`EphemerisSources.jl`](lib/EphemerisSources)   | A top-level package which provides, and re-exports, names from `HorizonsAPI`, `HorizonsEphemeris`, `SPICEKernels`, and `SPICEBodies`. |
+| [`HorizonsAPI.jl`](lib/HorizonsAPI)             | A precise JPL Horizons REST API client implementation, with keyword arguments for each acceptable parameter.                          |
+| [`HorizonsEphemeris.jl`](lib/HorizonsEphemeris) | Convenience wrappers around the JPL Horizons REST API.                                                                                |
+| [`SPICEKernels.jl`](lib/SPICEKernels)           | All [generic kernels](https://naif.jpl.nasa.gov/naif/data_generic.html) exported as variable constants.                               |
[…]
