### 2024-08-03 JuliaAstro/EphemerisSources.jl#15 @jdc-pub
The `(::KernelBody)(...)` method now outputs a named tuple, which is directly compatible with `StaticArrays.FieldVector` sub-types, `DataFrames`, `CSV`, and more! I have noticed that `NamedTuple` types are really performant when constructing `FieldVector` types, as I do in [`AstrodynamicalModels.jl`](https://github.com/cadojo/GeneralAstrodynamics.jl/blob/c2f5def232ca511bcae8a241e16044690db2d13e/lib/AstrodynamicalModels/src/AstrodynamicalModels.jl#L143).
