### 2024-08-03 #6914 @cadojo
Thank you all for the detailed review! I've been working a bit behind the scenes, moving this project to the @JuliaAstro organization, releasing `EphemerisSources.jl` and `EphemerisSourcesBase.jl`, and other structural changes. With those done, I should be able to finish the first round of review comments in the next week :) thank you for your patience and guidance here!
