### 2024-08-03 5c4c5ff
Resets brand image to JuliaAstro
