+++ lib/EphemerisSourcesBase/docs/src/index.md
@@ -11,12 +11,3 @@ _Common types and methods for `EphemerisSources.jl`!_
 
 This package is sparse... for now. By providing this lightweight dependency, 
 future versions of `EphemerisSources.jl` can provide more robust interfaces.
-
-# Reference
-
-```@autodocs
-Modules = [
-    EphemerisSourcesBase,
-]
-Order = [:module, :type, :function, :constant]
-```
