+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"reference/#Documentation","page":"Reference","title":"Documentation","text":"","category":"section"},{"location":"reference/","page":"Reference","title":"Reference","text":"All docstrings!","category":"page"},{"location":"reference/","page":"Reference","title":"Reference","text":"Modules = [\n    HorizonsAPI,\n]\nOrder = [:module, :type, :function, :constant]","category":"page"},{"location":"reference/#HorizonsAPI.HorizonsAPI","page":"Reference","title":"HorizonsAPI.HorizonsAPI","text":"Interact with the JPL HORIZONS REST API.\n\nExtended Help\n\nLicense\n\nMIT License\n\nCopyright (c) 2023 Joe Carpinelli\n\nPermission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the \"Software\"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:\n\nThe above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.\n\nTHE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.\n\nExports\n\nfetch_approach\nfetch_elements\nfetch_observer\nfetch_properties\nfetch_spk\nfetch_vectors\n\nImports\n\nBase\nCore\nDocStringExtensions\n\n\n\n\n\n","category":"module"},{"location":"reference/#HorizonsAPI.fetch_approach-Tuple{Any}","page":"Reference","title":"HorizonsAPI.fetch_approach","text":"fetch_approach(\n    COMMAND;\n    file,\n    CALIM_PL,\n    CALIM_SB,\n    CA_TABLE_TYPE,\n    EMAIL_ADDR,\n    TCA3SG_LIMIT,\n    format\n)\n\n\nFetch planetary ephemeris for the body specified by COMMAND, in the EPHEM_TYPE=\"APPROACH\" format.\n\n\n\n\n\n","category":"method"},{"location":"reference/#HorizonsAPI.fetch_elements-Tuple{Any}","page":"Reference","title":"HorizonsAPI.fetch_elements","text":"fetch_elements(\n    COMMAND;\n    file,\n    A,\n    EC,\n    ECLIP,\n    EMAIL_ADDR,\n    EPOCH,\n    IN,\n    MA,\n    N,\n    OBJECT,\n    OM,\n    QR,\n    TP,\n    W,\n    format\n)\n\n\nFetch planetary ephemeris for the body specified by COMMAND, in the EPHEM_TYPE=\"ELEMENTS\" format.\n\n\n\n\n\n","category":"method"},{"location":"reference/#HorizonsAPI.fetch_observer-Tuple{Any}","page":"Reference","title":"HorizonsAPI.fetch_observer","text":"fetch_o
[…]
