### 2024-08-03 83c853b
CompatHelper: add new compat entry for EphemerisSourcesBase at version 0.1, (keep existing compat)
