+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"reference/#Reference","page":"Reference","title":"Reference","text":"","category":"section"},{"location":"reference/","page":"Reference","title":"Reference","text":"All docstrings!","category":"page"},{"location":"reference/","page":"Reference","title":"Reference","text":"Modules = [\n    SPICEBodies,\n]\nOrder = [:module, :type, :function, :constant]","category":"page"},{"location":"reference/#SPICEBodies.SPICEBodies","page":"Reference","title":"SPICEBodies.SPICEBodies","text":"Ephemeris via SPICE kernels.\n\nExtended Help\n\nImports\n\nAstroTime\nBase\nCore\nDocStringExtensions\nLinearAlgebra\nSPICE\n\nExports\n\nKernelBody\ngm\nradii\n\n\n\n\n\n","category":"module"},{"location":"reference/#SPICEBodies.AbstractKernelBody","page":"Reference","title":"SPICEBodies.AbstractKernelBody","text":"abstract type AbstractKernelBody\n\nA supertype for all SPICE ephemeris objects.\n\nInterface\n\nAll subtypes must implement the following method.\n\nSPICEBodies.naifcode(body)::Int\n\nThe type then has access to the following method implementations.\n\nbody(epoch, ...)     # full state vector (KM, KM/S)\ngm(body)             # \"GM\" mass parameter (KM^3/S^2)\nradii(body)          # radii vector (KM)\n\n\n\n\n\n","category":"type"},{"location":"reference/#SPICEBodies.KernelBody","page":"Reference","title":"SPICEBodies.KernelBody","text":"struct KernelBody <: SPICEBodies.AbstractKernelBody\n\nAny celestial body, spacecraft, barycenter, or other ephemeris object which has an  assigned NAIF ID within the local SPICE kernel pool.\n\n\n\n\n\n","category":"type"},{"location":"reference/#SPICEBodies.gm-Tuple{Union{AbstractString, Integer, Symbol, SPICEBodies.AbstractKernelBody}}","page":"Reference","title":"SPICEBodies.gm","text":"gm(body)\n\n\nReturn the mass parameter of the body (KM^3/S^2).\n\n\n\n\n\n","category":"method"},{"location":"reference/#SPICEBodies.j2000s","page":"Reference","title":"SPICEBodies.j2000s","text":"Return the number of seconds since J2000.\n\n\n\n\n\n","category":"function"},{"location":"reference/#SPICEBodies.naifcode","page":"Reference","title":"SPICEBodies.naifcode","text":"Given an ephemeris body, a NAIF ID code, or a body name, return the corresponding NAIF ID.\n\n\n\n\n\n","category":"function"},{"location":"reference/#SPICEBodies.radii-Tuple{Union{AbstractString, Integer, Symbol, SPICEBodies.AbstractKernelBody}}","page":"Reference","title":"SPICEBodies.radii","text":"radii(body)\n\n\nReturn the radii vector of the body (KM).\n\n\n\n\n\n","category":"method"},{"location":"reference/#SPICEBodies.radius-Tuple{Union{AbstractString, Integer, Symbol, SPICEBodies.AbstractKernelBody}}","page":"Reference","title":"SPICEBodies.radius","text":"radius(body)\n\n\nReturn the two-norm of the radius vector of the body (KM).\n\n\n\n\n\n","category":"method"},{"location":"reference/#SPICEBodies.BodyLike","page":"Reference","title":"SPICEBodies.BodyLike","text":"A union type for all SPICEBodies.naifcode argument types.\n\n\n\n\n\n","category":"type"},{"location":"reference/#SPICEBodies.BodyLike-Tuple{Union{Real, 
[…]
