### 2024-08-03 1d0925c
Shows supported NAIF IDs in documentation
