+++ lib/HorizonsEphemeris/docs/Project.toml
@@ -5,3 +5,4 @@ DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 HorizonsAPI = "c15253bb-5e94-4b8b-9a02-579bb6c8e3ce"
 HorizonsEphemeris = "05ee1981-f730-42d8-b713-4f42d99733dc"
 LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
+SPICE = "5bab7191-041a-5c2e-a744-024b9c3a5062"
