+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"reference/#Reference","page":"Reference","title":"Reference","text":"","category":"section"},{"location":"reference/","page":"Reference","title":"Reference","text":"All docstrings!","category":"page"},{"location":"reference/","page":"Reference","title":"Reference","text":"Modules = [\n    HorizonsEphemeris,\n]\nOrder = [:module, :type, :function, :constant]","category":"page"},{"location":"reference/#HorizonsEphemeris.HorizonsEphemeris","page":"Reference","title":"HorizonsEphemeris.HorizonsEphemeris","text":"Interact with the JPL Horizons REST API.\n\nExtended Help\n\nLicense\n\nMIT License\n\nCopyright (c) 2023 Joe Carpinelli\n\nPermission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the \"Software\"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:\n\nThe above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.\n\nTHE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.\n\nExports\n\nNAIF\nephemeris\n\nImports\n\nBase\nCSV\nCore\nDates\nDocStringExtensions\nPrintf\n\n\n\n\n\n","category":"module"},{"location":"reference/#HorizonsEphemeris.MJD-Tuple{Any}","page":"Reference","title":"HorizonsEphemeris.MJD","text":"MJD(timestamp)\n\n\nGiven any timestamp, return the corresponding modified Julian date.\n\n\n\n\n\n","category":"method"},{"location":"reference/#HorizonsEphemeris.NAIF-Tuple{AbstractString}","page":"Reference","title":"HorizonsEphemeris.NAIF","text":"NAIF(name)\n\n\nGiven a NAIF ID, return the associated name, if one exists. If the ID provided cannot be found, a KeyError is thrown.\n\n\n\n\n\n","category":"method"},{"location":"reference/#HorizonsEphemeris.NAIF-Tuple{Integer}","page":"Reference","title":"HorizonsEphemeris.NAIF","text":"NAIF(code)\n\n\nIf a name is given, return the associated NAIF ID, if one exists. If the name provided cannot be found, a KeyError is thrown.\n\n\n\n\n\n","category":"method"},{"location":"reference/#HorizonsEphemeris.ephemeris-NTuple{4, Any}","page":"Reference","title":"HorizonsEphemeris.ephemeris","text":"ephemeris(\n    body,\n    start,\n    stop,\n    intervol;\n    site,\n    wrt,\n    file,\n    units,\n    frame,\n    header\n)\n\n\nPull ephemeris data for the provided celestial body, or celestial system.\n\n\n\n\n\n","category":"method"},{"location":"refe
[…]
