### 2024-08-03 JuliaAstro/EphemerisSources.jl#14 @jdc-pub
I've fixed `AstroTime` support, changed the default units to `KM-S` to match Horizons (good catch!), and addressed the other comments in [`1d0925cef8`](https://github.com/JuliaAstro/EphemerisSources.jl/commit/1d0925cef86d4bb8dd064f9e1bcf8bae789391ef).
