### 2024-08-04 c510905
Initial refactor for HorizonsEphemeris
