+++ lib/HorizonsEphemeris/gen/make.jl
@@ -1,0 +1,113 @@
+#
+# Code Generation
+#
+
+"""
+Given the full observer quantities list, create variable-name-safe labels.
+
+!!! note
+    This function was generated with ChatGPT.
+"""
+function create_labels(description_list::AbstractVector{<:AbstractString})
+    labels = similar(description_list)
+    for index in eachindex(description_list)
+        description = description_list[index]
+
+        # Remove leading and trailing spaces and split by first space (to remove numbers and any punctuations)
+        clean_description = split(description, '.'; limit=2)[end]
+        
+        # Generate the variable-name-safe label by replacing non-alphanumeric characters with underscores
+        label = replace(clean_description, r"\W" => "_")
+        label = replace(label, r"_+" => "_")  # Replace multiple underscores with a single underscore
+        label = strip(label, '_')  # Remove leading and trailing underscores
+        
+        # Ensure the label is capitalized
+        label = titlecase(label)
+        
+        # Add the pair to the labels vector
+        labels[index] = label
+    end
+    return labels
+end
+
+OBSERVER_QUANTITIES = 
+    """
+        1. Astrometric RA & DEC
+      * 2. Apparent RA & DEC
+        3.   Rates; RA & DEC
+      * 4. Apparent AZ & EL
+        5.   Rates; AZ & EL
+        6. Satellite X & Y, position angle
+        7. Local apparent sidereal time
+        8. Airmass and Visual Magnitude Extinction
+        9. Visual magnitude & surface Brightness
+       10. Illuminated fraction
+       11. Defect of illumination
+       12. Satellite angle of separation/visibility code
+       13. Target angular diameter
+       14. Observer sub-longitude & sub-latitude
+       15. Sun sub-longitude & sub-latitude
+       16. Sub-Sun position angle & distance from disc center
+       17. North pole position angle & sistance from disc center
+       18. Heliocentric ecliptic longitude & latitude
+       19. Heliocentric range & range-rate
+       20. Observer range & range-rate
+       21. One-way down-leg light-time
+       22. Speed of target with respect to Sun & observer
+       23. Sun-Observer-Targ ELONGATION angle
+       24. Sun-Target-Observer ~PHASE angle
+       25. Target-Observer-Moon/Illumination%
+       26. Observer-Primary-Target angle
+       27. Position Angles; radius & -velocity
+       28. Orbit plane angle
+       29. Constellation Name
+       30. Delta-T (TDB - UT)
+     * 31. Observer-centered Earth ecliptic longitude & latitude
+       32. North pole RA & DEC
+       33. Galactic longitude and latitude
+       34. Local apparent SOLAR time
+       35. Earth->Site light-time
+     > 36. RA & DEC uncertainty
+     > 37. Plane-of-sky (POS) error ellipse
+     > 38. Plane-of-sky (POS) uncertainty (RSS)
+     > 39. Range & range-rate sigma
+     > 40. Doppler/delay sigmas
+       41. True anomaly angle
+     * 42. Local apparent hour angle
+       43. PHASE angle & bisector
+       44. Apparent target-centered longitude of Sun (L_s)
+     * 45. Inertial frame apparent RA & DEC
+       46.   Rates: Inertial RA & DEC
+     * 47. Sky motion: angular rate & angles
+       48. Lunar sky brightness & target visual SNR
+    """
+
+open(joinpath(@__DIR__, "..", "src", "gen", "observer_quantities.jl"), "w") do file
+
+    quantities = create_labels(split(OBSERVER_QUANTITIES, "\n"))
+
+    write(
+        file, 
+        """
+        #
+        # This file is autogenerated!
+        #
+
+        """
+    )
+
[…]
