+++ lib/HorizonsEphemeris/src/HorizonsEphemeris.jl
@@ -39,7 +39,9 @@ using Dates
 using AstroTime
 using Printf
 using SPICE: bodc2n, bodn2c
-using HorizonsAPI: fetch_vectors
+using HorizonsAPI
+
+include(joinpath("gen", "observer_quantities.jl"))
 
 """
 The Horizons API can only process so many discrete time points at once!
@@ -124,68 +126,15 @@ MJD(timestamp::AbstractFloat) = timestamp - 2400000.5 < zero(timestamp) ? timest
 MJD(timestamp::Tuple) = MJD(collect(timestamp))
 MJD(timestamp::AbstractVector) = map(MJD, timestamp)
 
-"""
-Pull ephemeris data for the provided celestial body, or celestial system.
-"""
-function ephemeris(
-    body, start, stop, intervol;
-    site="",
-    wrt="ssb",
-    file=nothing,
-    units="KM-S",
-    frame="ICRF",
-    header=[:t, :cal, :x, :y, :z, :ẋ, :ẏ, :ż],
-)
-    code = body isa Integer ? body : NAIF(body)
+# fetch_properties, fetch_observer, fetch_vectors, fetch_spk, fetch_elements, fetch_approach
 
-    if !(uppercase(strip(units)) in ("AU-D", "KM-D", "KM-S"))
-        error("The only acceptable inputs for the units keyword argument are: KM-S, KM-D, AU-D.")
-    end
-
-    response = fetch_vectors(
-        code;
-        format="json",
-        CENTER="$(site)@$(NAIF(wrt))",
-        START_TIME=Dates.format(DateTime(start), "yyyy-mm-dd HH:MM:SS.sss"),
-        STOP_TIME=Dates.format(DateTime(stop), "yyyy-mm-dd HH:MM:SS.sss"),
-        STEP_SIZE=string(intervol),
-        REF_SYSTEM=string(frame),
-        CSV_FORMAT=true,
-        VEC_TABLE=2,
-        VEC_CORR="NONE",
-        OUT_UNITS=units,
-        VEC_LABELS=false,
-        VEC_DELTA_T=false,
-        TIME_DIGITS="FRACSEC"
-    )
-
-    ephemeris, notes = parse_response(response; start="\$\$SOE", stop="\$\$EOE")
-
-    csv = CSV.File(IOBuffer(ephemeris); header=false, drop=[9])
-
-    output = NamedTuple(label => csv[column] for (label, column) in zip(header, csv.names))
-
-    if !isnothing(file)
-        _, cal, _, _, _, _, _, _ = csv.names
-        text = NamedTuple(
-            label => column != cal ? map(f -> @sprintf("%f", f), csv[column]) : csv[column]
-            for (label, column) in zip(header, csv.names)
-        )
-
-        CSV.write(file, text; header=header)
-        @info "Ephemeris data for object with NAIF ID $code has been written to $file."
-    end
-
-    return output
-end
-
-function ephemeris(
+function HorizonsAPI.fetch_observer(
     body, times::Base.AbstractVecOrTuple;
     site="",
-    wrt="ssb",
+    wrt="earth",
     mjd=true,
     file=nothing,
-    units="AU-D",
+    units="KM-S",
     frame="ICRF",
     header=[:t, :cal, :x, :y, :z, :ẋ, :ẏ, :ż]
 )
@@ -279,13 +228,169 @@ function ephemeris(
 
 end
 
-function ephemeris(
+function HorizonsAPI.fetch_observer(
+    body, time;
+    site="",
+    wrt="earth",
+    mjd=true,
+    file=nothing,
+    units="KM-S",
+    frame="ICRF",
+)
+
+    code = body isa Integer ? body : NAIF(body)
+
+    if !(uppercase(strip(units)) in ("AU-D", "KM-D", "KM-S"))
[…]

[…]
