### 2024-08-04 45d8022
Switches default format to text in HorizonsAPI
