### 2024-08-04 JuliaAstro/EphemerisSources.jl#12 @jdc-pub
>  Why is SPICEApplications.jl mentioned in the paper, seems to be in the call stack, but is not part of this meta repo?

I originally had `SPICEApplications.jl` in this project, but I have recently moved it to `SPICE.jl`. I feel it's more appropriate there.
