+++ paper/paper.md
@@ -29,110 +29,139 @@ bibliography: references.bib
 
 # Summary
 
-Students and professionals in astronomy, astrodynamics, astrophysics, and other
-related fields often download and parse data about objects in our solar system
---- ephemeris data --- from two major providers:
-JPL's publicly available [Generic SPICE Kernels](https://naif.jpl.nasa.gov/pub/naif/generic_kernels/),
-and JPL's [Horizons platform](https://ssd.jpl.nasa.gov/horizons/). SPICE kernels are
-typically read through the SPICE Toolkit, which is available in a variety of
-programming languages, include the C Programming Language with `CSPICE` [@cspice].
-The Julia packages [`CSPICE_jll.jl`](https://github.com/JuliaBinaryWrappers/CSPICE_jll.jl)
-and [`SPICE.jl`](https://github.com/JuliaAstro/SPICE.jl) expose many `CSPICE`
-functions through Julia functions. Julia users can load and interact with SPICE
-kernels through methods such as `SPICE.furnsh` and `SPICE.spkez`.
-Horizons data is available through
-a variety of methods, including email, command-line, graphical web interfaces,
-and a [REST API](https://ssd-api.jpl.nasa.gov/doc/horizons.html) [@horizons].
+Students and professionals in astronomy, astrodynamics, astrophysics,
+and other related fields often download and parse data about objects in
+our solar system --- ephemeris data --- from two major providers: JPL's
+publicly available [Generic SPICE
+Kernels](https://naif.jpl.nasa.gov/pub/naif/generic_kernels/), and JPL's
+[Horizons platform](https://ssd.jpl.nasa.gov/horizons/). SPICE kernels
+are typically read through the SPICE Toolkit, which is available in a
+variety of programming languages, include the C Programming Language
+with `CSPICE` [@cspice]. The Julia packages
+[`CSPICE_jll.jl`](https://github.com/JuliaBinaryWrappers/CSPICE_jll.jl)
+and [`SPICE.jl`](https://github.com/JuliaAstro/SPICE.jl) expose many
+`CSPICE` functions through Julia functions. Julia users can load and
+interact with SPICE kernels through methods such as `SPICE.furnsh` and
+`SPICE.spkez`. Horizons data is available through a variety of methods,
+including email, command-line, graphical web interfaces, and a [REST
+API](https://ssd-api.jpl.nasa.gov/doc/horizons.html) [@horizons].
 
-This paper introduces several packages --- `SPICEKernels.jl`, `SPICEBodies.jl`,
-`HorizonsAPI.jl` and `HorizonsEphemeris.jl` --- which allow users to download and process
-ephemeris data idiomatically, all from within Julia.
-Through the use of these packages, users can share replicatable code which
-automatically fetches publicly available ephemeris data, as opposed to manually
+This paper introduces several packages --- `SPICEKernels.jl`,
+`SPICEBodies.jl`, `HorizonsAPI.jl` and `HorizonsEphemeris.jl` --- which
+allow users to download and process Cartesian state vector data
+idiomatically, all from within Julia. While ephemeris data comes in many
+forms, including observer tables, osculating orbital elements, and
+binary formats, these packages currently target Cartesian state vector
+(position and velocity) ephemeris data. Through the use of these
+packages, users can share replicatable code which automatically fetches
+data from publicly available ephemeris sources, as opposed to manually
 including ephemeris data files with their source code distribution.
 
 # Statement of Need
 
-While ephemeris users have the tools they need to fetch and parse ephemeris
-data from multiple sources within Julia, they do not have the tools to do so
-_simply_ or _idiomatically_. Horizons ephemeris data is distributed in plain text with
[…]
