### 2024-08-04 JuliaAstro/EphemerisSources.jl#12 @jdc-pub
I have addressed all paper comments in [`5dbdbbe450`](5dbdbbe450a327da147831c72bda118a11756ed7). This project now lives in the @JuliaAstro organization!
