### 2024-08-04 23ab74c
Create CODE_OF_CONDUCT.md
