### 2024-08-09 f8ac1a0
Updates register CI
