+++ .github/workflows/Register.yml
@@ -4,13 +4,32 @@ on:
   workflow_dispatch:
 
 jobs:
-  release:
-    name: Register Package
+  release-subpackages:
+    name: Register Subpackages
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        group:
+          - EphemerisSourcesBase
+          - HorizonsAPI
+          - HorizonsEphemeris
+          - SPICEKernels
+          - SPICEBodies
+    steps:
+      - uses: peter-evans/commit-comment@v3
+        with:
+          body: |
+            @JuliaRegistrator register subdir="lib/${{ matrix.group }}"
+
+            This comment was generated with `commit-comment` and [`Register.yml`](/.github/workflows/Register.yml).
+
+  release-superpackage:
+    name: Register Superpackage
     runs-on: ubuntu-latest
     steps:
       - uses: peter-evans/commit-comment@v3
         with:
           body: |
-            @JuliaRegistrator register subdir="lib/SPICEKernels"
+            @JuliaRegistrator register"
 
             This comment was generated with `commit-comment` and [`Register.yml`](/.github/workflows/Register.yml).
