### 2024-08-09 1c2ac73
Removes explicit versions in docs
