+++ lib/HorizonsAPI/docs/make.jl
@@ -18,5 +18,4 @@ deploydocs(
     branch = "horizons-api",
     devbranch = "main",
     tag_prefix = "HorizonsAPI-",
-    versions = ["stable" => "HorizonsAPI-v^", "HorizonsAPI-v#.#", "HorizonsAPI-v#.#.#"],
 )
