+++ .github/workflows/TagBot.yml
@@ -1,5 +1,4 @@
 name: TagBot
-
 on:
   issue_comment:
     types:
@@ -8,7 +7,6 @@ on:
     inputs:
       lookback:
         default: "3"
-
 permissions:
   actions: read
   checks: read
@@ -22,39 +20,12 @@ permissions:
   repository-projects: read
   security-events: read
   statuses: read
-
 jobs:
   TagBot:
     if: github.event_name == 'workflow_dispatch' || github.actor == 'JuliaTagBot'
     runs-on: ubuntu-latest
     steps:
-      - name: Tag EphemerisSourcesBase.jl
-        uses: JuliaRegistries/TagBot@v1
+      - uses: JuliaRegistries/TagBot@v1
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
           ssh: ${{ secrets.DOCUMENTER_KEY }}
-          subdir: lib/EphemerisSourcesBase
-      - name: Tag HorizonsAPI.jl
-        uses: JuliaRegistries/TagBot@v1
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          ssh: ${{ secrets.DOCUMENTER_KEY }}
-          subdir: lib/HorizonsAPI
-      - name: Tag HorizonsEphemeris.jl
-        uses: JuliaRegistries/TagBot@v1
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          ssh: ${{ secrets.DOCUMENTER_KEY }}
-          subdir: lib/HorizonsEphemeris
-      - name: Tag SPICEKernels.jl
-        uses: JuliaRegistries/TagBot@v1
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          ssh: ${{ secrets.DOCUMENTER_KEY }}
-          subdir: lib/SPICEKernels
-      - name: Tag SPICEBodies.jl
-        uses: JuliaRegistries/TagBot@v1
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          ssh: ${{ secrets.DOCUMENTER_KEY }}
-          subdir: lib/SPICEBodies
