+++ dev/lib/HorizonsEphemeris/versions.js
@@ -1,2 +1,7 @@
 var DOC_VERSIONS = [
+  "stable",
+  "v1.0",
+  "dev",
 ];
+var DOCUMENTER_NEWEST = "v1.0.1";
+var DOCUMENTER_STABLE = "stable";
