+++ Project.toml
@@ -1,7 +1,7 @@
 name = "EphemerisSources"
 uuid = "858900df-f7ae-402a-aa8c-3e4ef85aa271"
 authors = ["Joey Carpinelli <joseph.d.carpinelli@loopy.codes>"]
-version = "0.1.0"
+version = "0.1.1"
 
 [deps]
 AstroTime = "c61b5328-d09d-5e37-a9a8-0eb41c39009c"
