### 2024-08-09 #6914 @cadojo
> Please ping here when you want us to take a look again to confirm things. I tried taking a quick peek at #13, which you closed, and some of those items marked as completed still seemed to be problems present on main. However, I am not really sure how you are versioning atm.

I'll ping you when the repository is ready for review. I do think I addressed all items, but there have been some issues getting the repository's documentation to host correctly. I think I've sorted that out tonight. I will soon address the final issue you left, and I will ping y'all here when everything is ready for re-review. Thanks again --- I know this is taking longer than expected, but I'm learning a lot about Julia tooling nuances after all of these structural and ownership changes.
