+++ v0.1.1/lib/SPICEKernels/versions.js
@@ -1,7 +1,7 @@
 var DOC_VERSIONS = [
   "stable",
-  "v0.1",
+  "v1.4",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.1.1";
+var DOCUMENTER_NEWEST = "v1.4.0";
 var DOCUMENTER_STABLE = "stable";
