### 2024-08-15 #6914 @cadojo
@andrew-saydjari @tamasgal The issues in the repository are all correctly updated. I have a couple of items left in the single open issue, and those are unchecked. I'm in the process of doing those now, but otherwise the repository is ready for any more feedback you all have. I have sorted out the documentation hosting issues, CI issues, etc.

Could someone please update the repository to https://github.com/JuliaAstro/EphemerisSources.jl?
