### 2024-09-04 #6914 @adonath
Thanks @andrew-saydjari for the note. It seems the editorial bot is happy with the paper source, however the last changes I see in the source markdown are from 4 weeks ago: https://github.com/JuliaAstro/EphemerisSources.jl/commits/main/paper/paper.md

@cadojo Can you confirm this is correct? Or are the changes to the paper still in the old repository?
