### 2024-09-06 JuliaAstro/EphemerisSources.jl#16 @jdc-pub
Velocity example posted here: https://juliaastro.org/EphemerisSources.jl/docs/dev/examples/velocity-study/
