### 2024-09-06 JuliaAstro/EphemerisSources.jl#13 @andrew-saydjari
Some initial comments
- [x] Broken link: https://github.com/JuliaAstro/HorizonsAPI.jl
- [x] Broken link: https://github.com/JuliaAstro/HorizonsEphemeris.jl
- [x] Broken link: https://github.com/JuliaAstro/SPICEBodies.jl
- [x] Broken link: https://github.com/JuliaAstro/SPICEKernels.jl
- [x] Plots in some examples (for instance https://juliaastro.org/EphemerisSources.jl/docs/dev/examples/spice/) are not rendering
- [x] Broken link: https://juliaastro.org/EphemerisSources.jl/lib/HorizonsAPI/stable
- [x] Broken link: https://juliaastro.org/EphemerisSources.jl/lib/SPICEKernels/stable
- [x] Case study would benefit from measured minus ideal, instead of the two separate plots, but that is just a preference.
