### 2024-09-09 JuliaAstro/EphemerisSources.jl#13 @andrew-saydjari
Everything seems fixed except

- [x]  You have the package name Horizons.jl on the https://github.com/cadojo/EphemerisSources.jl?tab=readme-ov-file page, but it appears to be HORIZONS.jl on its github (and in your paper). Probably should change it for consistency.
