### 2024-09-09 JuliaAstro/EphemerisSources.jl#14 @andrew-saydjari
It appears the example plot

plot(
   earth.x, earth.y;
   aspect_ratio = 1,
   linewidth = 1.5,
   border = :none,
   size = (600, 600),
   dpi = 200,
   grid = false,
   axis = nothing,
   title = "",
   label=:none,
   color = "green",
   background=:transparent,
)

Has been removed from HorizonsEphemeris.jl? Any reason why? I thought it was great!
