### 2024-09-09 JuliaAstro/EphemerisSources.jl#14 @andrew-saydjari
I cannot reproduce the AstroTime support, but maybe I am doing something wrong with the versioning. I am using [05ee1981] HorizonsEphemeris v1.2.1. I would have installed from git like I did with Pkg.add(url="https://github.com/JuliaAstro/EphemerisSources.jl") to get the latest version, but the subpackages (like HorizonsEphemeris) do not seem to have git pages that allow direct download.  Can you either update the registered package or tell me if my invocation is wrong?

```
ep = TTEpoch(59235.0days, origin=:modified_julian)

earth = let start = ep - 50years, stop = ep + 50years, step = 1days
   @time ephemeris("earth", start, stop, step; wrt="jupiter", units="AU-D");
    # what are my units choices, what are my named bodies options?
end
```

Everything else listed on this issue has been resolved.
