### 2024-09-09 JuliaAstro/EphemerisSources.jl#15 @andrew-saydjari
While trying to check this, I hit.

```
⌅ [8e9d28ce] SPICEKernels v1.4.4 (<v2.0.0): EphemerisSources
```

Seems like some compat to fix. I am not sure if your compat helper in the outer directory is managing the subpackages correctly.
