### 2024-09-09 JuliaAstro/EphemerisSources.jl#15 @andrew-saydjari
But I was able to successfully confirm that the named tuple output is working. So the above comment is really just to let you know about your compat management. My review of this issue is complete.
