### 2024-09-09 JuliaAstro/EphemerisSources.jl#16 @andrew-saydjari
Can you point me to where you added guidelines to address?
- [x] I did not see any community guidelines (which is one of our review checklist items). Please add something that addresses that requirement to help people know how willing you are to interact with them in terms of contributions, issues, and support.
