### 2024-09-10 JuliaAstro/EphemerisSources.jl#16 @jdc-pub
@andrew-saydjari I had thought the [code of conduct](https://github.com/JuliaAstro/EphemerisSources.jl/blob/main/CODE_OF_CONDUCT.md) was the standard "community guidelines" for projects on GitHub. I'm sure that file is not an issue, but to satisfy this item, should I add something else?
