### 2024-09-10 JuliaAstro/EphemerisSources.jl#14 @jdc-pub
Do you mean [this plot](https://juliaastro.org/EphemerisSources.jl/docs/dev/examples/horizons/#Reference) above the _References_ section? It's pretty hidden by that massive DataFrame. I've reduced the number of rows in the latest version of the docs.

Good catch, it seems `HorizonsEphmeris` had an issue where it allowed floating point types for the `interval`, but Horizons does not. I've corrected this. In the latest version of the project, I was able to run the code example you provided above.

As for the units and bodies choice, I don't know if Horizons supplies a full list of the bodies it supports. For the units, I updated the doctoring for `ephemeris` to clarify that units of KM-S, KM-D, and AU-D are supported.
