### 2024-09-10 JuliaAstro/EphemerisSources.jl#16 @andrew-saydjari
The exact wording I have is " Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" My view is that this is different from a code of conduct. Here you are telling the people reading your README.md how you feel about contributions. Some packages have such high use rates that developers cannot respond to every issue or will only respond to PRs passing tests/some formatting restrictions. Some packages are happy to take issues, PRs, bug reports, feature requests, etc. People generally put a sentence at the bottom of the readme to explain their stance on that. See some examples
- https://github.com/SciML/DifferentialEquations.jl, which directs some questions to Gitter, some to issues, and developers to a special part of the documentation.
- https://github.com/andrew-saydjari/CloudCovErr.jl for a minimal example from me... just telling people we know things are still preliminary and are committed to fixing it if they make an issue.

You can be anywhere on this spectrum of complexity of statement (and also be less open to contributions), but I think the goal of this item on the checklist is to make sure you have said *something* about your position on contributions/issues/etc.
