### 2024-09-10 JuliaAstro/EphemerisSources.jl#14 @andrew-saydjari
Yes. I did not scroll down past that huge table. You might just invert the order or trim the data frame to only a few lines (I still see 261 lines...). But ok, thanks for pointing out my error there. Style choice to you if you want to increase its visibility.

Sorry for the comment in my AstroTime example. I think you already had answered that question previously. That was just a leftover note to myself when I was doing the first review pass.

I am still not able to run the code example however.  [858900df] EphemerisSources v0.3.0 [https://github.com/JuliaAstro/EphemerisSources.jl.git#main](https://github.com/JuliaAstro/EphemerisSources.jl.git#main%60). This still gives me the error.

```
The data delimiters ($$SOE, $$EOE) were not found in the response!

	Unknown units specification -- re-enter
```

Part of me might be confused with how users *should* be invoking these subpackages. Should people really being calling Pkg.add("HorizonsEphemeris"), or should they do Pkg.add("EphemerisSources") and calling the subpackages as EphemerisSources.HorizonsEphemeris.ephemeris, for example? I tried the latter when trying this time, to ensure I had the most recent version of the code.
