### 2024-09-10 JuliaAstro/EphemerisSources.jl#15 @andrew-saydjari
Confirmed. Happy with the completion of this issue.
