### 2024-09-13 JuliaAstro/EphemerisSources.jl#14 @jdc-pub
> Part of me might be confused with how users should be invoking these subpackages. Should people really being calling Pkg.add("HorizonsEphemeris"), or should they do Pkg.add("EphemerisSources") and calling the subpackages as EphemerisSources.HorizonsEphemeris.ephemeris, for example? I tried the latter when trying this time, to ensure I had the most recent version of the code.

Users should be able to do either, but I had intended users to use individual sub-packages if they knew what they needed, and use `EphemerisSources.jl` if they want everything at once.

> I am still not able to run the code example however.

Sorry about that. I do think this is fixed now. I also added this as a [test](https://github.com/juliaastro/ephemerissources.jl/blob/main/lib/HorizonsEphemeris/test/runtests.jl) to `HorizonsEphemeris`, and it is passing. Could you please try one more time? Maybe the recent `HorizonsEphemeris-v1.3.0` version update fixed things.
