### 2024-09-13 JuliaAstro/EphemerisSources.jl#14 @andrew-saydjari
Interesting. It does appear fixed in HorizonEphemeris-v1.3.0 just using the ephemeris exported there. However, before doing that update, git installing EphemerisSources and using EphemerisSources.HorizonsEphemeris.ephemeris was still failing, but after doing the Pkg.update("HorizonEphemeris") EphemerisSources.HorizonsEphemeris.ephemeris started working. This suggests to me that there is some sort of dependency issue. Maybe it means that the Project.toml for EphemerisSources was being satisfied by the HorizonEphemeris v1.2.0 I had installed and so it was not updating it... but I would have thought a direct git install would have gotten the most recent/up to date HorizonEphemeris version directly through the subfoldering. Just calling this out because it did not work as I expected.

In any case, yes, I agree the functionality for AstroTime has been included. Just a note about verifying subpackage versioning.
