### 2024-09-14 #6914 @cadojo
All sounds great, thank you everyone! This was my first submission to JOSS, and I really appreciate everyone's patience and guidance.

@adonath please let me know if/when you have any edits/suggestions.
