### 2024-10-03 #6914 @adonath
Thanks @cadojo and sorry for the slight delay. I proof read the paper carefully and read very well. So I don't have any further comments on the text. However it would be good to add the DOIs of the references as indicated by the editorial bot. The reference "Status of the JPL Horizons Ephemeris System", seems to not have a DOI. So OK to keep as is.

Once the missing DOI has been added, please go ahead and release a new version of the software and post the version tag here. Please also archive the software (e.g. on Zenodo) and share the DOI. Thanks!
