### 2024-10-06 e004a1e
Paper typo fixes
