+++ paper/paper.md
@@ -2,10 +2,10 @@
 title: Idiomatic Ephemeris Sourcing and Parsing in Julia
 
 tags:
-  - astrodynamics
-  - astronomy
-  - astrophysics
-  - ephemeris
+  - Astrodynamics
+  - Astronomy
+  - Astrophysics
+  - Ephemeris
   - Julia
   - SPICE
   - Horizons
@@ -32,17 +32,17 @@ bibliography: references.bib
 Students and professionals in astronomy, astrodynamics, astrophysics,
 and other related fields often download and parse data about objects in
 our solar system --- ephemeris data --- from two major providers: JPL's
-publicly available [Generic SPICE
-Kernels](https://naif.jpl.nasa.gov/pub/naif/generic_kernels/), and JPL's
+publicly-available [Generic SPICE
+Kernels](https://naif.jpl.nasa.gov/pub/naif/generic_kernels/) and JPL's
 [Horizons platform](https://ssd.jpl.nasa.gov/horizons/). SPICE kernels
 are typically read through the SPICE Toolkit, which is available in a
-variety of programming languages, include the C Programming Language
+variety of programming languages, including the C Programming Language
 with `CSPICE` [@cspice]. The Julia packages
 [`CSPICE_jll.jl`](https://github.com/JuliaBinaryWrappers/CSPICE_jll.jl)
 and [`SPICE.jl`](https://github.com/JuliaAstro/SPICE.jl) expose many
 `CSPICE` functions through Julia functions. Julia users can load and
 interact with SPICE kernels through methods such as `SPICE.furnsh` and
-`SPICE.spkez`. Horizons data is available through a variety of methods,
+`SPICE.spkez`. Horizons provides data through a variety of methods,
 including email, command-line, graphical web interfaces, and a [REST
 API](https://ssd-api.jpl.nasa.gov/doc/horizons.html) [@horizons].
 
@@ -53,8 +53,8 @@ idiomatically, all from within Julia. While ephemeris data comes in many
 forms, including observer tables, osculating orbital elements, and
 binary formats, these packages currently target Cartesian state vector
 (position and velocity) ephemeris data. Through the use of these
-packages, users can share replicatable code which automatically fetches
-data from publicly available ephemeris sources, as opposed to manually
+packages, users can share replicable code which automatically fetches
+data from publicly-available ephemeris sources, as opposed to manually
 including ephemeris data files with their source code distribution.
 
 # Statement of Need
@@ -62,10 +62,10 @@ including ephemeris data files with their source code distribution.
 While astronomers, astrodynamicists, and other ephemeris users have the
 tools they need to fetch and parse position and velocity data from
 multiple sources within Julia, they do not have the tools to do so
-*simply* or *idiomatically*. Horizons ephemeris data is distributed in
+*simply* or *idiomatically*. Horizons' ephemeris data is distributed in
 plain text with surrounding metadata, and manual parsing is required for
 users to programmatically use the fetched ephemeris data. Generic SPICE
-kernels are freely available, and can be used with `CSPICE` (and wrapper
+kernels are freely available and can be used with `CSPICE` (and wrapper
 libraries) for kernel inspection and data retrieval, but new users and
 students may find the required workflows unfamiliar. The packages
 presented in this paper may be used by students and professionals to
@@ -83,13 +83,14 @@ REST API with tab-completion through *static keyword arguments*[^1], and
 the first to offer automatic response parsing into `NamedTuple` types.
 The `NamedTuple` output of `HorizonsEphemeris.ephemeris`, the top-level
 method for fetching Cartesian state vectors from the Horizons platform,
[…]
