### 2024-10-06 #6914 @cadojo
I have fixed the references, and published new versions of the software, and released the software as an [archive](https://zenodo.org/records/13895994) on Zenodo. The DOI for the project is [`10.5281/zenodo.13895994`](https://doi.org/10.5281/zenodo.13895994). The version tag for the super-package --- `EphemerisSources.jl` --- is [`v1.0.0`](https://github.com/JuliaAstro/EphemerisSources.jl/commit/e004a1e2c6bfd2b8e1afd3539201dc518f451df3). I expect `TagBot` will release a new tag for this version (and the new versions of the sub-packages) shortly.

Thank you all for your support and feedback!
