### 2024-10-08 #6914 @cadojo
@adonath I made some [very small](https://github.com/JuliaAstro/EphemerisSources.jl/commit/71f219b51c7f8d81ab50ca98a8dc4bac1932f661) edits, mostly to move those pesky final words (before references) back onto the second page. I'm going to @editorialbot generate pdf one more time, but if that slows down what's left of this publication process, feel free to disregard. Again, thank you for all of your time.
