### 2024-10-10 #6914 @warrickball
Almost there! As a subtle point in the [post-review checklist](https://github.com/openjournals/joss-reviews/issues/6914#issuecomment-2391385598), could you update the title of the [Zenodo repository](https://zenodo.org/records/13895994) to match the paper? I.e. it should also be "Idiomatic Ephemeris Sourcing and Parsing in Julia".

You can change the archive metadata without creating a new archive.

I'm aware this is inherited from the repo title, which is "EphemerisSources.jl". We don't need you to change that but you may want to align all the titles by changing them to something like "EphemerisSources.jl: Idiomatic Ephemeris Sourcing and Parsing in Julia".
