### 2024-10-10 #6914 @cadojo
Done! Both Zenodo and the paper's title have been updated.
