### 2024-10-10 #6914 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
