### 2024-10-10 #6914 @warrickball
Thanks once again to @tamasgal and @andrew-saydjari for reviewing and @adonath for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.

Congratulations @cadojo, your paper has been published in JOSS! :tada:
